% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add2vec}
\alias{add2vec}
\title{Add the contents of one vector to another vector}
\usage{
add2vec(target, increments, inds)
}
\arguments{
\item{dest}{target vector}

\item{orig}{vector to be added}

\item{indices}{in the target}
}
\value{
add the elements of one row vector to another.
}
\description{
Add the contents of one vector to another vector
}
