% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afss_settings.R
\name{afss_settings}
\alias{afss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
afss_settings(factor_update_interval = 1, weight_update_interval = NULL,
  n_fss_updates = NULL, adapt_factors = TRUE, initial_widths = NULL,
  initial_factors = NULL)
}
\arguments{
\item{factor_update_interval}{number of iterations between factor and
interval updates (defaults to an update every iteration)}

\item{weight_update_interval}{number of iterations between updates to the
factor sampling weights, defaults to the factor update interval.}

\item{n_fss_updates}{number of slice directions that are sampled per
iteration, defaults to the number of model parameters if NULL.}

\item{adapt_factors}{should the factors be updated? defaults to TRUE}

\item{initial_widths}{vector of initial slice widths}

\item{initial_factors}{matrix of initial factor loadings}

\item{adapt_intervals}{should the interval}
}
\value{
list with additional settings for automated factor slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
