% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks2cov.R
\name{blocks2cov}
\alias{blocks2cov}
\title{Reconstitute a joint kernel covariance matrix from block covariances}
\usage{
blocks2cov(kernel_objects, parameter_blocks)
}
\arguments{
\item{kernel_objects}{list of kernel objects containing the sub-block
covariance matrices}

\item{parameter_blocks}{list of parameter blocks}
}
\value{
reconstituted covariance matrix
}
\description{
Reconstitute a joint kernel covariance matrix from block covariances
}
