% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_flowmat.R
\name{build_flowmat}
\alias{build_flowmat}
\title{Given a list of (unparsed) rate functions, construct the matrix of
compartment flows.}
\usage{
build_flowmat(rates, compartment_names)
}
\arguments{
\item{rates}{Unparsed list of rate functions (intermediate object within the
\code{stem_dynamics} function)}

\item{compartment_names}{vector of compartment_strata names}
}
\value{
matrix whose elements indicate the change in each compartment
  (columns) associated with each reaction event (rows).
}
\description{
Given a list of (unparsed) rate functions, construct the matrix of
compartment flows.
}
