% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lna_fcns.R
\name{build_lna_fcns}
\alias{build_lna_fcns}
\title{Reformat the rates for use in the LNA and compute the derivatives of the rate
functions.}
\usage{
build_lna_fcns(rate_fcns, param_codes, const_codes, tcovar_codes,
  compartment_codes, flow_matrix, lna_scale)
}
\arguments{
\item{rate_fcns}{list of rates, as returned by stem_dynamics}

\item{param_codes}{vector of parameter codes}

\item{const_codes}{vector of time-varying covariate codes}

\item{tcovar_codes}{vector of constant codes}

\item{flow_matrix}{matrix indicating compartment flow, whose transpose is the
stoichiometry matrix}

\item{lna_scale}{either "log" or "linear"}

\item{comp_codes}{vector of compartment codes}
}
\value{
strings for computing the LNA hazards and the LNA jacobian
}
\description{
Reformat the rates for use in the LNA and compute the derivatives of the rate
functions.
}
