% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lna_rates.R
\name{build_lna_rates}
\alias{build_lna_rates}
\title{Reformat the rates for use in the LNA and compute the derivatives of the rate functions.}
\usage{
build_lna_rates(rates, param_codes, const_codes, tcovar_codes,
  compartment_codes)
}
\arguments{
\item{rates}{list of rates, as returned by stem_dynamics}

\item{param_codes}{vector of parameter codes}

\item{const_codes}{vector of time-varying covariate codes}

\item{tcovar_codes}{vector of constant codes}

\item{comp_codes}{vector of compartment codes}
}
\value{
character vector of rate derivatives ordered by compartment within
  rates
}
\description{
Reformat the rates for use in the LNA and compute the derivatives of the rate functions.
}

