% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_measproc_indmat.R
\name{build_measproc_indmat}
\alias{build_measproc_indmat}
\title{Construct an indicator matrix for which measurement process variables are
measured at which observation times.}
\usage{
build_measproc_indmat(obsmat)
}
\arguments{
\item{obsmat}{observation matrix with NA indicating no observation at a
particular time.}
}
\value{
logical matrix indicating measurment status
}
\description{
Construct an indicator matrix for which measurement process variables are
measured at which observation times.
}
