% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_obsmat.R
\name{build_obsmat}
\alias{build_obsmat}
\title{Generate a template for a stemr observation matrix, or combine multiple
datasets into a stemr observation matrix; used internally.}
\usage{
build_obsmat(meas_procs = NULL, datasets = NULL)
}
\arguments{
\item{meas_procs}{parsed list of emission lists, generated internally within
\code{\link{stem_measure}}.}

\item{datasets}{list of datasets.}
}
\value{
observation matrix
}
\description{
Generate a template for a stemr observation matrix, or combine multiple
datasets into a stemr observation matrix; used internally.
}

