% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ode_rates.R
\name{build_ode_rates}
\alias{build_ode_rates}
\title{Construct the ODE hazard functions}
\usage{
build_ode_rates(rate_fcns, param_codes, const_codes, tcovar_codes,
  compartment_codes)
}
\value{
list of rate function strings and parameter codes
}
\description{
Construct the ODE hazard functions
}

