% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rate_adjmat.R
\name{build_rate_adjmat}
\alias{build_rate_adjmat}
\title{Construct an adjacency matrix for specifying which rates must be updated when
a transition occurs, with adjacency determined at the lumped population
level.}
\usage{
build_rate_adjmat(rates, compartment_codes)
}
\arguments{
\item{rates}{intermediate list of rate functions created within the
stem_dynamics function}

\item{compartment_codes}{vector of compartment codes.}
}
\value{
adjacency matrix
}
\description{
Constructs an adjacency matrix with rates given in the rows, and reactions on
which they each depend given in columns. Thus, when a reaction occurs, all of
the non-zero entries in the corresponding column must be updated.
}
