% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tcovar_adjmat.R
\name{build_tcovar_adjmat}
\alias{build_tcovar_adjmat}
\title{Construct an adjacency matrix for which rates need to be update when there is
a change in the time-varying covariates.}
\usage{
build_tcovar_adjmat(rates, tcovar_codes = NULL)
}
\arguments{
\item{rates}{list of rates in ste_dynamics}

\item{tcovar_codes}{character vector of time-varying covariate names}
}
\value{
adjacency matrix for which rates need to be updated when a
  time-varying covariate changes.
}
\description{
Constructs an adjacency matrix with rates given in the rows, time, and
time-varying covariates on which they each depend given in columns. Thus,
when time or a time-varying covariate changes, all of the non-zero entries in
the corresponding column must be updated.
}
