% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tcovar_changemat.R
\name{build_tcovar_changemat}
\alias{build_tcovar_changemat}
\title{Indicator matrix for determining which time-varying covariates change at each
row of a time-varying covariate matrix.}
\usage{
build_tcovar_changemat(tcovar)
}
\arguments{
\item{tcovar}{time-varying covariate matrix}
}
\value{
indicator matrix for which time-varying covariates change at teach
  time in a time-varying covariate matrix.
}
\description{
Indicator matrix for determining which time-varying covariates change at each
row of a time-varying covariate matrix.
}

