% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_rw}
\alias{c_rw}
\title{Componentwise Metropolis random walk transition kernel}
\usage{
c_rw(params_prop, params_cur, ind, kernel_cov)
}
\arguments{
\item{params_prop}{vector in which the proposed parameters should be stored}

\item{params_cur}{vector containing the current parameter vector}

\item{ind}{C++ style index for the component index}

\item{kernel_cov}{vector of component proposal standard deviations}
}
\value{
propose new parameter values in place
}
\description{
Componentwise Metropolis random walk transition kernel
}
