% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_initdist_loglik.R
\name{calc_initdist_loglik}
\alias{calc_initdist_loglik}
\title{Return the log-lik of initial distribution draws}
\usage{
calc_initdist_loglik(initdist_objects)
}
\arguments{
\item{initdist_objects}{list of initial distribution objects}
}
\value{
log likelihood
}
\description{
Return the log-lik of initial distribution draws
}
