% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_param_logprior.R
\name{calc_params_logprior}
\alias{calc_params_logprior}
\title{Return the log prior density over all parameter blocks}
\usage{
calc_params_logprior(param_blocks)
}
\arguments{
\item{param_blocks}{list of parameter blocks}
}
\value{
log prior density
}
\description{
Return the log prior density over all parameter blocks
}
