% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_COMPUTE_LNA}
\alias{CALL_COMPUTE_LNA}
\title{Compute the LNA by calling the LNA functions via external Xptr.}
\usage{
CALL_COMPUTE_LNA(t, state, parms)
}
\arguments{
\item{t}{time}

\item{state}{vector containing the drift, residual, and diffusion subvectors}

\item{parms}{list containing the LNA model parameters, stoichiometry matrix,
and LNA pointer}
}
\description{
Compute the LNA by calling the LNA functions via external Xptr.
}

