% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_D_MEASURE}
\alias{CALL_D_MEASURE}
\title{Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.}
\usage{
CALL_D_MEASURE(
  emitmat,
  emit_inds,
  record_ind,
  record,
  state,
  parameters,
  constants,
  tcovar,
  d_meas_ptr
)
}
\arguments{
\item{emitmat}{matrix of emission probabilities}

\item{emit_inds}{logical vector of measurement compartments to simulate}

\item{record_ind}{row in the observation and emission matrices}

\item{record}{vector of observed counts}

\item{state}{numeric vector of latent comaprtment counts}

\item{parameters}{numeric vector of parameter values}

\item{constants}{numeric vector of constants}

\item{tcovar}{numeric vector of time-varying covariate values}

\item{d_meas_ptr}{external pointer to measurement process simulation function}
}
\description{
Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.
}
