% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_INTEGRATE_STEM_LNA}
\alias{CALL_INTEGRATE_STEM_LNA}
\title{Compute the hazards by calling the hazard functions via external Xptr.}
\usage{
CALL_INTEGRATE_STEM_LNA(init, start, end, step_size, lna_ode_ptr)
}
\arguments{
\item{init}{initial condition}

\item{end}{time at right endpoint}

\item{step_size}{set automatically by caller, required argument not specified by user}

\item{lna_ode_ptr}{external pointer for calling the ODE integrator}

\item{state}{time at left endpoint of interval}
}
\description{
Compute the hazards by calling the hazard functions via external Xptr.
}
