% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_LNA_ESS}
\alias{CALL_LNA_ESS}
\title{Compute the LNA by calling the LNA functions via external Xptr.}
\usage{
CALL_LNA_ESS(t, state, parms)
}
\arguments{
\item{t}{time}

\item{state}{vector of compartment counts}

\item{list}{containing the vector of model parameters, stoichiometry matrix,
and elliptical slice sampling LNA pointer}
}
\description{
Compute the LNA by calling the LNA functions via external Xptr.
}

