% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_R_MEASURE}
\alias{CALL_R_MEASURE}
\title{Simulate from the measurement process by calling measurement process
functions via external Xptr.}
\usage{
CALL_R_MEASURE(obsmat, emit_inds, record_ind, state, parameters, constants,
  tcovar, r_meas_ptr)
}
\arguments{
\item{obsmat}{observation matrix}

\item{emit_inds}{logical vector of measurement compartments to simulate}

\item{record_ind}{row in the observation matrix}

\item{state}{numeric vector of comaprtment counts}

\item{parameters}{numeric vector of parameter values}

\item{constants}{numeric vector of constants}

\item{tcovar}{numeric vector of time-varying covariate values}

\item{r_meas_ptr}{external pointer to measurement process simulation function}
}
\description{
Simulate from the measurement process by calling measurement process
functions via external Xptr.
}
