% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_RATE_FCN}
\alias{CALL_RATE_FCN}
\title{Update rates by calling rate functions via Xptr.}
\usage{
CALL_RATE_FCN(rates, inds, state, parameters, constants, tcovar, rate_ptr)
}
\arguments{
\item{rates}{vector of rates to be modified}

\item{inds}{logical vector of indices of rates to be modified}

\item{state}{numeric vector of comaprtment counts}

\item{parameters}{numeric vector of parameter values}

\item{constants}{numeric vector of constants}

\item{tcovar}{numeric vector of time-varying covariate values}

\item{rate_ptr}{external pointer to rate function}
}
\description{
Update rates by calling rate functions via Xptr.
}
