% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_SET_LNA_PARAMS}
\alias{CALL_SET_LNA_PARAMS}
\title{Compute the hazards by calling the hazard functions via external Xptr.}
\usage{
CALL_SET_LNA_PARAMS(p, set_lna_params_ptr)
}
\arguments{
\item{t}{time}

\item{state}{vector of compartment counts}

\item{parameters}{vector of model parameters}

\item{constants}{vector of constants}

\item{vector}{of time-varying covariate values}

\item{hazard_ptr}{external pointer to function used to compute the hazards}
}
\description{
Compute the hazards by calling the hazard functions via external Xptr.
}

