% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_SET_ODE_PARAMS}
\alias{CALL_SET_ODE_PARAMS}
\title{Set the parameters for a system of ODEs via XPtr.}
\usage{
CALL_SET_ODE_PARAMS(p, set_ode_params_ptr)
}
\arguments{
\item{p}{vector of parameters}

\item{set_ode_params_ptr}{external pointer to the ODE parameter setting function.}
}
\description{
Set the parameters for a system of ODEs via XPtr.
}
