% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{census_incidence}
\alias{census_incidence}
\title{Construct a matrix containing the incidence counts at a sequence of census times.}
\usage{
census_incidence(incid_mat, census_times, interval_inds)
}
\arguments{
\item{incid_mat}{matrix with incidence counts}

\item{census_times}{vector of census times}

\item{interval_inds}{interval indices, generated by a call to findInterval with left.open=T}
}
\value{
matrix containing the incidence counts at census times.
}
\description{
Construct a matrix containing the incidence counts at a sequence of census times.
}
