% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{census_lna}
\alias{census_lna}
\title{Construct a matrix containing the compartment counts and the incidence at a sequence of census times.}
\usage{
census_lna(path, census_path, census_inds, lna_event_inds, flow_matrix_lna,
  do_prevalence, init_state, lna_pars, forcing_inds, forcing_tcov_inds,
  forcings_out, forcing_transfers)
}
\arguments{
\item{path}{matrix containing the path to be censused (cumulative incidence).}

\item{census_path}{matrix to be filled out with the path.}

\item{census_inds}{vector of indices for census interval endpoints.}

\item{lna_event_inds}{vector of column indices in the path matrix for events that
should be censused.}

\item{flow_matrix_lna}{matrix containing the flow matrix for the LNA (no incidence)}

\item{do_prevalence}{should the prevalence be computed}

\item{init_state}{the initial compartment counts}

\item{lna_pars}{matrix with parameters, constants, and time varying 
covariates and parameters.}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Construct a matrix containing the compartment counts and the incidence at a sequence of census times.
}
