% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_path.R
\name{census_path}
\alias{census_path}
\title{Wrapper for evaluating the compartment counts at census times.}
\usage{
census_path(path, census_times, census_columns = NULL, census_matrix = NULL)
}
\arguments{
\item{path}{matrix containing the path to be censused.}

\item{census_times}{vector of census times}

\item{census_columns}{indices of columns to be censused. If not specified,
will default to all columns except 'time', 'event', and 'ID'.}

\item{census_matrix}{existing matrix of compartment counts at census times to
be modified in place. If not specified, a new matrix will be returned.}
}
\value{
matrix of compartment counts at census times
}
\description{
Wrapper for evaluating the compartment counts at census times.
}

