% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_path_collection.R
\name{census_path_collection}
\alias{census_path_collection}
\title{Census each path in a collection of paths to obtain the compartment counts at
census times.}
\usage{
census_path_collection(paths, census_times, census_columns = NULL,
  as_array = FALSE)
}
\arguments{
\item{paths}{either a list or array of paths}

\item{census_times}{vector of census times}

\item{census_columns}{indices of columns to be censused. If not specified,
will default to all columns except 'time', 'event', and 'ID'.}

\item{as_array}{return the matrices of compartment counts as an array?}
}
\value{
a list (default) or array of matrices of compartment counts at census
  times.
}
\description{
Census each path in a collection of paths to obtain the compartment counts at
census times.
}

