% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tpar_depends.R
\name{check_tpar_depends}
\alias{check_tpar_depends}
\title{Check if time varying parameters depend on initial conditions}
\usage{
check_tpar_depends(tparam, initdist_objects, parmat)
}
\arguments{
\item{tparam}{list of time varying parameters}

\item{initdist_objects}{list with initial distribution objects}

\item{parmat}{parameter matrix}
}
\value{
time-varying parameter list with indicators for whether each
  parameter depends on initial conditions
}
\description{
Check if time varying parameters depend on initial conditions
}
