% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comp_sqrtmat}
\alias{comp_sqrtmat}
\title{matrix square root}
\usage{
comp_sqrtmat(S, M, nugget)
}
\arguments{
\item{S}{square root matrix to be filled out}

\item{M}{symmetric positive definite matrix for which square root is to be computed}

\item{nugget}{small positive constant to be added to the diagonal for numerical stability}
}
\value{
set S equal to the matrix square root of M
}
\description{
matrix square root
}
