% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_lna_ess.R
\name{compile_lna_ess}
\alias{compile_lna_ess}
\title{Instatiate the C++ rate functions for the LNA elliptical slice sampling in
which only the residual and drift processes need to be integrated.}
\usage{
compile_lna_ess(lna_rates, flow_matrix, lna_scale, messages = TRUE)
}
\arguments{
\item{lna_rates}{list containing the strings vectors of ODES for the hazards
and jacobian matrix, along with the lna parameter codes.}

\item{flow_matrix}{matrix indicating the changes to each compartment per
event}

\item{lna_scale}{either "log" or "linear"}

\item{messages}{logical; print a message that the rates are being compiled}
}
\value{
External function pointers for integrating the LNA
}
\description{
Instatiate the C++ rate functions for the LNA elliptical slice sampling in
which only the residual and drift processes need to be integrated.
}

