% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_incidence}
\alias{compute_incidence}
\title{Difference an incidence variable in a census matrix.}
\usage{
compute_incidence(censusmat, col_inds, row_inds)
}
\arguments{
\item{censusmat}{matrix of compartment counts at census times, to be updated}

\item{col_inds}{column indices for which incidence should be computed}

\item{row_inds}{list of vectors of row indices for each of the incidence variables}
}
\value{
update the census matrix in place
}
\description{
Difference an incidence variable in a census matrix.
}
