% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lna_density.R
\name{compute_lna_density}
\alias{compute_lna_density}
\title{Compute the log-transition density for a latent LNA path}
\usage{
compute_lna_density(path, fixed_inits, log_scale, incidence_codes)
}
\arguments{
\item{path}{list containing the current path and the paths of its ODEs}

\item{fixed_inits}{is the initial state vector fixed?}

\item{log_scale}{is the LNA applied to the log transformed process?}

\item{incidence_codes}{vector of incidence codes}
}
\value{
log density of the sampled LNA path
}
\description{
Compute the log-transition density for a latent LNA path
}

