% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lna_density2.R
\name{compute_lna_density2}
\alias{compute_lna_density2}
\title{Compute the log density of an LNA path by integrating the LNA ODEs.}
\usage{
compute_lna_density2(path, parameters, stoich_matrix, lna_pointer, times,
  param_update_inds, drift_inds, resid_inds, diff_inds, log_scale,
  incidence_codes, fixed_inits)
}
\arguments{
\item{path}{matrix containing a sampled LNA path}

\item{parameters}{list of lna parameters, the lna function pointer, and the
stoichiometry matrix}

\item{stoich_matrix}{the stoichiometry matrix, i.e. the transpose of the flow
matrix.}

\item{lna_pointer}{external pointer to the LNA function}

\item{times}{vector of times at which the LNA is to be evaluated}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{drift_inds}{vector of indices corresponding to the drift process within
the vector holding the ODEs}

\item{resid_inds}{vector of indices corresponding to the residual process
within the vector holding the ODEs}

\item{diff_inds}{vector of indices corresponding to the diffusion process
within the vector holding the ODEs}

\item{log_scale}{logical}

\item{incidence_codes}{vector of incidence codes}

\item{fixed_inits}{is the initial state fixed}
}
\value{
log-density of the LNA path
}
\description{
Compute the log density of an LNA path by integrating the LNA ODEs.
}

