% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lna_logliks.R
\name{compute_lna_logliks}
\alias{compute_lna_logliks}
\title{Compute the LNA log likelihood and data log likelihood.}
\usage{
compute_lna_logliks(path, data, params_prop_nat, lna_params_prop, censusmat,
  emitmat, flow_matrix, lna_times, lna_initdist_inds, param_update_inds,
  incidence_codes, census_incidence_codes, census_indices, measproc_indmat,
  obstime_inds, constants, tcovar_censmat, do_prevalence, do_incidence,
  lna_pointer, lna_set_pars_pointer, d_meas_pointer)
}
\arguments{
\item{path}{list with the current LNA path along with its ODE paths to be
reintegrated}

\item{data}{matrix containing the dataset}

\item{params_prop_nat}{vector of proposed parameters on their natural scale}

\item{lna_params_prop}{matrix with proposed LNA parameters}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{lna_times}{times at whicht eh LNA should be evaluated}

\item{lna_initdist_inds}{C++ column indices in the LNA parameter matrix for
the initial state}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{do_prevalence}{should prevalence be computed?}

\item{lna_pointer}{external LNA pointer}

\item{lna_set_pars_pointer}{pointer for setting the LNA parameters}

\item{d_meas_pointer}{external pointer for the measurement process function}
}
\value{
list with an updated LNA path along with its ODEs, the observed data
  log-likelihood, and the lna log-likelihood.
}
\description{
Compute the LNA log likelihood and data log likelihood.
}
