% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_initdist_sampler.R
\name{construct_initdist_sampler}
\alias{construct_initdist_sampler}
\title{Construct a function for sampling new initial compartment counts, which are
taken to have multinomial prior.}
\usage{
construct_initdist_sampler(state_initializer, n_strata, constants)
}
\arguments{
\item{state_initializer}{list for intializing the comparment counts generated
by \code{\link{stem_dynamics}}.}

\item{n_strata}{number of strata.}

\item{constants}{vector of model constants, which contains the population
size and strata sizes.}
}
\value{
function for evaluating the log prior density of the initial
  compartment counts
}
\description{
Construct a function for sampling new initial compartment counts, which are
taken to have multinomial prior.
}

