% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{construct_priors}
\alias{construct_priors}
\title{Construct a list of functions for evaluating prior densities of model
parameters.}
\usage{
construct_priors(priors, param_codes)
}
\arguments{
\item{priors}{list of parameters prior lists, each constructed via a call to
the \code{\link{prior}} function.}

\item{param_codes}{vector of parameter codes as generated by the
\code{\link{stem_dynamics}} function}
}
\value{
function to evaluate the prior densities, returning a vector
}
\description{
Construct a list of functions for evaluating prior densities of model
parameters.
}

