% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_lna}
\alias{convert_lna}
\title{Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale.}
\usage{
convert_lna(path, flow_matrix, init_state, log_scale)
}
\arguments{
\item{path}{matrix containing the LNA path in terms of the counting
processes on transition events}

\item{flow_matrix}{stoichiometry matrix (the transpose of the flow matrix)}

\item{init_state}{initial compartment counts on the natural scale}

\item{log_scale}{was the lna path computed on the log scale?

The process can be re-expressed by left-multiplying each row in the path
matrix by the stoichiometry matrix: \eqn{X_t = A'\phi(t, N_t)}.}
}
\description{
Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale.
}

