% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_lna2}
\alias{convert_lna2}
\title{Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale, making the conversion in place
for an existing census matrix.}
\usage{
convert_lna2(path, flow_matrix, init_state, statemat)
}
\arguments{
\item{path}{matrix containing the LNA path in terms of the counting
processes on transition events}

\item{flow_matrix}{stoichiometry matrix (the transpose of the flow matrix)}

\item{init_state}{initial compartment counts on the natural scale}

\item{statemat}{matrix where the compartment counts should be written

The process can be re-expressed by left-multiplying each row in the path
matrix by the stoichiometry matrix: \eqn{X_t = A'\phi(t, N_t)}.}
}
\description{
Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale, making the conversion in place
for an existing census matrix.
}
