% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_2_rows}
\alias{copy_2_rows}
\title{Copy some of the rows of one matrix into another}
\usage{
copy_2_rows(dest, orig, inds)
}
\arguments{
\item{dest}{destination matrix}

\item{orig}{origin matrix}

\item{inds}{row indices}
}
\value{
copy the elements of one matrix into another.
}
\description{
Copy some of the rows of one matrix into another
}
