% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_elem}
\alias{copy_elem}
\title{Copy an element from one vector into another}
\usage{
copy_elem(dest, orig, ind)
}
\arguments{
\item{dest}{destination row vector}

\item{orig}{origin row vector}

\item{ind}{C++ style index for the element to be copied}
}
\value{
copy an element of one row vector into another.
}
\description{
Copy an element from one vector into another
}
