% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_mat}
\alias{copy_mat}
\title{Copy the contents of one matrix into another}
\usage{
copy_mat(dest, orig)
}
\arguments{
\item{dest}{destination row vector}

\item{orig}{origin row vector}
}
\value{
copy the elements of one matrix into another.
}
\description{
Copy the contents of one matrix into another
}
