% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_pathmat}
\alias{copy_pathmat}
\title{Copy the columns of one matrix into another}
\usage{
copy_pathmat(dest, orig)
}
\arguments{
\item{dest}{destination matrix}

\item{orig}{origin matrix}

\item{ind}{column index}
}
\value{
copy the elements of one matrix into another.
}
\description{
Copy the columns of one matrix into another
}
