% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_row}
\alias{copy_row}
\title{Copy the contents of one matrix into another}
\usage{
copy_row(dest, orig, ind)
}
\arguments{
\item{dest}{destination matrix}

\item{orig}{origin matrix}

\item{ind}{row index}
}
\value{
copy the elements of one matrix into another.
}
\description{
Copy the contents of one matrix into another
}
