% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_vec}
\alias{copy_vec}
\title{Copy the contents of one vector into another}
\usage{
copy_vec(dest, orig)
}
\arguments{
\item{dest}{destination row vector}

\item{orig}{origin row vector}
}
\value{
copy the elements of one row vector into another.
}
\description{
Copy the contents of one vector into another
}

