% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_vec2}
\alias{copy_vec2}
\title{Copy the contents of one vector into another}
\usage{
copy_vec2(dest, orig, inds)
}
\arguments{
\item{dest}{destination row vector}

\item{orig}{origin row vector}

\item{inds}{vector of indices in the destination}
}
\value{
copy the elements of one row vector into another.
}
\description{
Copy the contents of one vector into another
}
