% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copypars}
\alias{copypars}
\title{Identify which rates to update when a state transition event occurs.}
\usage{
copypars(dest, orig)
}
\arguments{
\item{dest}{destination row vector}

\item{orig}{origin row vector}
}
\value{
copy the elements of one row vector into another.
}
\description{
Identify which rates to update when a state transition event occurs.
}

