% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvtn}
\alias{dmvtn}
\title{Multivariate normal density}
\usage{
dmvtn(x, mu, sigma, logd = FALSE)
}
\arguments{
\item{x}{matrix of draws for which to evaluate the density}

\item{mu}{mean vector of the distribution}

\item{sigma}{covariance matrix}

\item{logd}{should the log be returned

(source: http://gallery.rcpp.org/articles/dmvnorm_arma/)}
}
\description{
Multivariate normal density
}
