% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doEdgeR.R
\name{doEdgeR}
\alias{doEdgeR}
\title{Function performs edgeR two-sample alculations}
\usage{
doEdgeR(counts, group)
}
\arguments{
\item{counts}{numeric matrix of read counts.}

\item{group}{vector or factor giving the experimental group/condition for each sample/library.}
}
\description{
Function performs edgeR two-sample alculations
}

