% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doFiltering.R
\name{doFiltering}
\alias{doFiltering}
\title{Function performs filtering, removes rows with low counts}
\usage{
doFiltering(counts, avg.min = 1)
}
\arguments{
\item{counts}{numeric matrix of read counts.}

\item{avg.min}{minimum average per row.}
}
\description{
Function performs filtering, removes rows with low counts
}

