% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_normals}
\alias{draw_normals}
\title{Draw new N(0,1) values and fill a vector.}
\usage{
draw_normals(v)
}
\arguments{
\item{v}{vector to fill with new N(0,1) draws}
}
\value{
draw new values in place
}
\description{
Draw new N(0,1) values and fill a vector.
}
