% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_normals2}
\alias{draw_normals2}
\title{Draw new N(0,1) values and fill a matrix.}
\usage{
draw_normals2(M)
}
\arguments{
\item{M}{matrix to fill with new N(0,1) draws}
}
\value{
draw new values in place
}
\description{
Draw new N(0,1) values and fill a matrix.
}
