% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_d_measure}
\alias{evaluate_d_measure}
\title{Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.}
\usage{
evaluate_d_measure(emitmat, obsmat, statemat, measproc_indmat, parameters,
  constants, tcovar_censusmat, d_meas_ptr)
}
\arguments{
\item{emitmat}{matrix of emission probabilities}

\item{parameters}{numeric vector of parameter values}

\item{constants}{numeric vector of constants}

\item{emit_inds}{logical vector of measurement compartments to simulate}

\item{record_ind}{row in the observation and emission matrices}

\item{record}{vector of observed counts}

\item{state}{numeric vector of latent comaprtment counts}

\item{tcovar}{numeric vector of time-varying covariate values}

\item{r_meas_ptr}{external pointer to measurement process simulation function}
}
\description{
Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.
}

