% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_d_measure_LNA}
\alias{evaluate_d_measure_LNA}
\title{Evaluate the log-density of a possibly time-verying measurement process
by calling measurement process density functions via external Xptr.}
\usage{
evaluate_d_measure_LNA(
  emitmat,
  obsmat,
  censusmat,
  measproc_indmat,
  parameters,
  param_inds,
  const_inds,
  tcovar_inds,
  param_update_inds,
  census_indices,
  param_vec,
  d_meas_ptr
)
}
\arguments{
\item{emitmat}{matrix of emission probabilities}

\item{obsmat}{matrix containing the data}

\item{censusmat}{matrix containing the state of the latent process at
observation times}

\item{measproc_indmat}{logical matrix indicating which compartments are
observed at every observation time}

\item{parameters}{matrix containing the LNA parameters, constants and
time-varying coariates.}

\item{param_inds}{indices for the model parameters.}

\item{const_inds}{indices for the constants.}

\item{tcovar_inds}{indices for the time-varying covariates.}

\item{param_update_inds}{logical vector indicating when model parameters
should be updated.}

\item{census_indices}{vector of indices when the LNA path has been censused.}

\item{param_vec}{vector for keeping the current lna parameters}

\item{d_meas_ptr}{external pointer to measurement process density function}
}
\description{
Evaluate the log-density of a possibly time-verying measurement process
by calling measurement process density functions via external Xptr.
}
