% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_interval}
\alias{find_interval}
\title{Given a vector of interval endpoints \code{breaks}, determine in which
intervals the elements of a vector \code{x} fall.}
\usage{
find_interval(x, breaks, rightmost_closed, all_inside)
}
\arguments{
\item{x}{vector for whose elements the corresponding intervals are
identified}

\item{breaks}{vector containing the elements}

\item{rightmost_closed}{logical; if true, the results for x[j]=breaks[N] is
N-1.}

\item{all_inside}{logical; if true, 0 is mapped to 1, and N is mapped to N-1

The rightmost interval is assumed to be closed. Compares to the behavior of
the \code{findInterval} function in \code{R}, when \code{rightmost.closed =
 TRUE}}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Given a vector of interval endpoints \code{breaks}, determine in which
intervals the elements of a vector \code{x} fall.
}

