% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stem.R
\name{fit_stem}
\alias{fit_stem}
\title{Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.}
\usage{
fit_stem(
  stem_object,
  method,
  priors,
  mcmc_kern,
  iterations,
  lna_ess_args = NULL,
  tparam_ess_args = NULL,
  thin_params = ceiling(iterations/1000),
  thin_latent_proc = ceiling(iterations/1000),
  initialization_attempts = 500,
  print_progress = 0,
  status_filename = NULL,
  messages = FALSE
)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "lna" or ode.}

\item{priors}{A list of three functions supplied by the user with names
"prior_density", "to_estimation_scale", and "from_estimation_scale" (N.B.
All three must be supplied). The first of these functions should have two
arguments that are numeric vectors of parameters on their natural and
estimation scales, specified in that order. The functions for converting
parameters to and from their estimation scales. IMPORTANT NOTES: 1) The
scale conversion functions must include the dots argument, "...", in the
function arguments. 2) The priors should not include priors for the initial
compartment counts or t0, which are specified in the
\code{stem_initializer} function and in the \code{t0_kernel} argument.}

\item{mcmc_kern}{MCMC transition kernel generated by a call to the
\code{mcmc_kernel} function.}

\item{iterations}{number of iterations}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100) so that every 100th path will be saved}

\item{initialization_attempts}{number of initialization attempts}

\item{print_progress}{interval at which to print progress to a text file. If
0 (default) progress is not printed.}

\item{status_filename}{string to pre-append to status files, defaults to LNA
or ODE depending on the method used.}

\item{messages}{should status messages be printed? defaults to FALSE.}

\item{ess_args}{list of elliptical slice sampling arguments}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.
}
