% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{forcing}
\alias{forcing}
\title{Declare a time varying covariate to be a forcing variable that moves
individuals between model compartments at discrete times.}
\usage{
forcing(tcovar_name, from, to)
}
\arguments{
\item{tcovar_name}{name of the time varying covariate}

\item{from}{name of the compartment that individuals exit}

\item{to}{name of the compartment that individuals enter}
}
\value{
list containing the forcing specification
}
\description{
Declare a time varying covariate to be a forcing variable that moves
individuals between model compartments at discrete times.
}
\examples{
forcings = list(forcing("Su2Sv", "S_unvaccinated", "S_vaccinated"))
}
