% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_estimation_scale}
\alias{from_estimation_scale}
\title{Transform the parameters from estimation scales to their natural scales}
\usage{
from_estimation_scale(scaled_params, scales)
}
\arguments{
\item{scaled_params}{numeric vector of scaled parameter values}

\item{scales}{character vector of estimation scales, either 'linear', 'log',
or 'logit'}

\item{natural_params}{numeric vector of parameter values on their natural
scale, to be modified in place}
}
\value{
updates the natural_params vector in place with the inverse of the
  scale transformation functions
}
\description{
Transform the parameters from estimation scales to their natural scales
}
