% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g_prop2c_prop}
\alias{g_prop2c_prop}
\title{Get componentwise proposals from a global proposal.}
\usage{
g_prop2c_prop(g2c_mat, params_cur, params_prop)
}
\arguments{
\item{g2c_mat}{matrix in which to keep the componentwise proposals}

\item{params_cur}{vector containing the current parameter vector}

\item{params_prop}{vector in which the proposed parameters}
}
\value{
fill g2c_mat with componentwise proposals
}
\description{
Get componentwise proposals from a global proposal.
}
