% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rw3.R
\name{generate_rw3}
\alias{generate_rw3}
\title{Generate objects for setting up a random walk of order 3}
\usage{
generate_rw3(ntimes)
}
\arguments{
\item{ntimes}{number of times (nodes)}
}
\value{
list containing the difference matrix, the structure matrix, the
  singular value decomposition of the structure matrix, and the reference
  standard deviation of the random walk (i.e., the geometric mean standard
  deviation of the random walk with precision 1).
}
\description{
Generate objects for setting up a random walk of order 3
}
