% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stan_code.R
\name{generate_stan_code}
\alias{generate_stan_code}
\title{Generate the user defined functions for fitting the restarting,
log-transformed LNA for SEM counting processes via a noncentered
parameterization in Stan.}
\usage{
generate_stan_code(dynamics, measurement_process, separate_params = TRUE,
  out.file = NULL)
}
\arguments{
\item{dynamics}{compiled stemr dynamics}

\item{measurement_process}{compiled stemr measurement process}

\item{separate_params}{should the model parameters be separated out (default) in the
template or combined (e.g. for multivariate normal) for convenience?}

\item{out.file}{optional file to output code.}
}
\description{
Generate the user defined functions for fitting the restarting,
log-transformed LNA for SEM counting processes via a noncentered
parameterization in Stan.
}
