% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_census_path}
\alias{get_census_path}
\title{Obtain the compartment counts at a sequence of census times.}
\usage{
get_census_path(path, census_times, census_columns)
}
\arguments{
\item{path}{matrix containing the path to be censused.}

\item{census_times}{vector of census times.}

\item{census_columns}{vector of column indices to be censused (C++ indexing
beginning at 0).}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Obtain the compartment counts at a sequence of census times.
}

