% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lna_rates.R
\name{get_lna_rates}
\alias{get_lna_rates}
\title{Reformat the rates for use in the LNA and compute the derivatives of the rate functions.}
\usage{
get_lna_rates(rates, parameters, tcovar, constants, compartments)
}
\arguments{
\item{rates}{list of rates, as returned by stem_dynamics}

\item{parameters}{vector of parameter codes}

\item{tcovar}{vector of time-varying covariate codes}

\item{constants}{vector of constant codes}

\item{compartments}{vector of compartment codes}
}
\value{
character vector of rate derivatives ordered by compartment within
  rates
}
\description{
Reformat the rates for use in the LNA and compute the derivatives of the rate functions.
}

