% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harss_settings.R
\name{harss_settings}
\alias{harss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
harss_settings(n_harss_updates = 1, bracket_update_interval = 100,
  bracket_limits = c(0, Inf))
}
\arguments{
\item{n_harss_updates}{number of hit-and-run updates.}

\item{bracket_update_interval}{how often the bracket update interval should be updated}

\item{bracket_limits}{limits for the slice bracket width, defaults to (0,Inf).}
}
\value{
list with additional settings for hit and run slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
