% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_and_run_slice_sampler_ode.R
\name{hit_and_run_slice_sampler_ode}
\alias{hit_and_run_slice_sampler_ode}
\title{Update model parameters via factor slice sampling}
\usage{
hit_and_run_slice_sampler_ode(model_params_est, model_params_nat,
  params_prop_est, params_prop_nat, har_direction, harss_bracket_width,
  n_expansions_harss, n_contractions_harss, n_harss_updates, path, data, priors,
  params_logprior_cur, ode_params_cur, ode_param_vec, tparam, censusmat,
  emitmat, flow_matrix, stoich_matrix, ode_times, forcing_inds, forcing_matrix,
  ode_param_inds, ode_const_inds, ode_tcovar_inds, ode_initdist_inds,
  param_update_inds, ode_event_inds, census_indices, measproc_indmat, svd_sqrt,
  svd_d, svd_U, svd_V, ode_pointer, ode_set_pars_pointer, d_meas_pointer,
  do_prevalence, step_size)
}
\arguments{
\item{model_params_est}{vector of model parameters on the estimation scale}

\item{model_params_nat}{vector of model parameters on the natural scale}

\item{params_prop_est}{vector for proposed model parameters on their
estimation scale}

\item{params_prop_nat}{vector for proposed parameters on their natural scale}

\item{har_direction}{vector for the hit and run direction}

\item{harss_bracket_width}{width of the bracket}

\item{n_expansions_harss}{vector for number of hit and run expansions}

\item{n_contractions_harss}{vector for number of hit and run contractions}

\item{n_harss_updates}{number of hit and run updates}

\item{path}{list containing the ode path and likelihood}

\item{data}{matrix containing the data}

\item{priors}{list with functions for computing the prior density and
transformations to and from the estimation scale}

\item{params_logprior_cur}{log prior density of model parameters}

\item{ode_params_cur}{matrix with current ode parameters, tcovar, tparam,
etc.}

\item{ode_param_vec}{vector for ode parameters}

\item{tparam}{list with time-varying parameters}

\item{censusmat}{matrix for storing the ode path at census times}

\item{emitmat}{matrix for emission probabilities}

\item{flow_matrix}{flow matrix}

\item{stoich_matrix}{stoichiometry matrix}

\item{ode_times}{times at which the ode is evaluated}

\item{forcing_inds}{indices at which forcings are applied}

\item{forcing_matrix}{matrix containing forcings}

\item{ode_param_inds}{indices for ode parameters for computing emission probs}

\item{ode_const_inds}{indices for constants used in computing emission probs}

\item{ode_tcovar_inds}{indices for time-varying covariates}

\item{ode_initdist_inds}{index for where the initial compartment volumes
begin}

\item{param_update_inds}{indices for when ode parameters should be updated}

\item{ode_event_inds}{codes for elementary events}

\item{census_indices}{indices for when the ode path should be censused}

\item{measproc_indmat}{matrix giving the indices for data likelihood
contributions}

\item{svd_sqrt}{matrix for computing the matrix square root for the ode}

\item{svd_d}{vector for ode singular values}

\item{svd_U}{matrix for ode left singular vectors}

\item{svd_V}{matrix for ode right singular vectors}

\item{ode_pointer}{external pointer for ode}

\item{ode_set_pars_pointer}{external pointer for setting ode parameters}

\item{d_meas_pointer}{external pointer for computing emission probabilities}

\item{do_prevalence}{should prevalence be computed}

\item{step_size}{initial step size for ODE stepper}
}
\value{
update the model parameters, path, and likelihood terms in place
}
\description{
Update model parameters via factor slice sampling
}
