% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence2prevalence.R
\name{incidence2prevalence}
\alias{incidence2prevalence}
\title{Convert an LNA/ODE incidence path to a prevalence path.}
\usage{
incidence2prevalence(path, flow_matrix, init_state, forcings = NULL,
  forcing_matrix = NULL)
}
\arguments{
\item{path}{matrix containing the incidence path the first column of which
are incidence times}

\item{flow_matrix}{flow matrix}

\item{init_state}{vector of initial compartment counts}

\item{forcings}{list of forcing lists}

\item{forcing_matrix}{matrix containing the forcings the first column of
which are times of forcings}
}
\value{
matrix containing the prevalence path
}
\description{
Convert an LNA/ODE incidence path to a prevalence path.
}
