% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{increment_elem}
\alias{increment_elem}
\title{Increment an element of a vector by 1}
\usage{
increment_elem(vec, ind)
}
\arguments{
\item{vec}{destination row vector}

\item{ind}{C++ style index for the element to be copied}
}
\value{
Add 1 to an element of a vector
}
\description{
Increment an element of a vector by 1
}
