% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{increment_vec}
\alias{increment_vec}
\title{Add one vector to another}
\usage{
increment_vec(target, increments)
}
\arguments{
\item{dest}{target vector}

\item{orig}{vector to be added}
}
\value{
add the elements of one row vector to another.
}
\description{
Add one vector to another
}
