% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_lna.R
\name{initialize_lna}
\alias{initialize_lna}
\title{Initialize the latent LNA path}
\usage{
initialize_lna(lna_parameters, flow_matrix, lna_pointer, lna_set_pars_pointer,
  lna_times, fixed_inits, param_update_inds, incidence_codes, data,
  measproc_indmat, obstime_inds, d_meas_pointer, parameters, constants,
  tcovar_censmat, initialization_attempts)
}
\arguments{
\item{flow_matrix}{original flow matrix giving the changes to compartments
from each reaction}

\item{lna_pointer}{external pointer to LNA integration fcn}

\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{data}{matrix containing the data}

\item{measproc_indmat}{matrix indicating which compartments are measured}

\item{obstime_inds}{list of indices indicating at which rows each measured
variable is recorded}

\item{d_meas_pointer}{external pointer to density function}

\item{parameters}{vector of model parameters}

\item{constants}{vector of constants}

\item{tcovar_censmat}{time-varying covariates at observation times}

\item{initialization_attempts}{maximum number of initialization attempts}

\item{incidence_codes_lna}{vector of incidence codes}
}
\value{
list containing an initial latent path along with the ODEs
}
\description{
Initialize the latent LNA path
}

