% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_ode.R
\name{initialize_ode}
\alias{initialize_ode}
\title{Initialize the ODE path}
\usage{
initialize_ode(
  dat,
  parmat,
  param_blocks,
  tparam,
  censusmat,
  emitmat,
  stoich_matrix,
  proc_pointer,
  set_pars_pointer,
  census_times,
  param_vec,
  param_inds,
  const_inds,
  tcovar_inds,
  initdist_inds,
  param_update_inds,
  census_indices,
  event_inds,
  measproc_indmat,
  d_meas_pointer,
  do_prevalence,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  initialization_attempts,
  step_size,
  initdist_objects
)
}
\arguments{
\item{parmat}{matrix with parameters, contants, time-varying pars and covars}

\item{param_blocks}{list of parameter blocks}

\item{tparam}{list of time-varying parameters}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{ODE stoichiometry matrix}

\item{proc_pointer}{external LNA pointer}

\item{set_pars_pointer}{pointer for setting the LNA parameters}

\item{census_times}{times at which the LNA should be evaluated}

\item{param_vec}{vector for storing lna parameters when evaluating the
measurement process}

\item{param_inds}{C++ column indices for parameters}

\item{const_inds}{C++ column indices for constants}

\item{tcovar_inds}{C++ column indices for time varying covariates}

\item{initdist_inds}{C++ column indices in the LNA parameter matrix for
the initial state}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{event_inds}{vector of column indices in the LNA path for which
incidence will be computed.}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{initialization_attempts}{number of initialization attempts}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{data}{matrix containing the dataset}

\item{par_init_fcn}{function for initializing the parameter values}
}
\value{
ODE path
}
\description{
Initialize the ODE path
}
