% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insert_block}
\alias{insert_block}
\title{Insert one matrix into another}
\usage{
insert_block(dest, orig, rowinds, colinds)
}
\arguments{
\item{dest}{destination matrix}

\item{orig}{origin matrix}

\item{rowinds}{vector of row indices}

\item{colinds}{vector of column indices}
}
\value{
copy the elements of one matrix into another.
}
\description{
Insert one matrix into another
}
