% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insert_elem}
\alias{insert_elem}
\title{insert an element into a vector}
\usage{
insert_elem(dest, elem, ind)
}
\arguments{
\item{dest}{destination row vector}

\item{ind}{C++ style index for the element to be copied}

\item{orig}{elem}
}
\value{
copy an element of one row vector into another.
}
\description{
insert an element into a vector
}
