% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_initdists.R
\name{insert_initdist}
\alias{insert_initdist}
\title{Insert natural scale parameters into a parameter matrix}
\usage{
insert_initdist(parmat, initdist_objects, prop)
}
\arguments{
\item{parmat}{matrix into which parameters should be inserted}

\item{initdist_objects}{list of parameter blocks}

\item{prop}{should proposed parameters be inserted?}
}
\value{
insert natural scale parameters into a parameter matrix
}
\description{
Insert natural scale parameters into a parameter matrix
}
