% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insert_lna_step}
\alias{insert_lna_step}
\title{Insert the lna step into the path matrix}
\usage{
insert_lna_step(path, lna_step, ind)
}
\arguments{
\item{path}{matrix containing the full LNA path}

\item{lna_step}{vector containing the sampled states}

\item{ind}{row index}
}
\value{
modifies the path matrix in place
}
\description{
Insert the lna step into the path matrix
}

