% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insert_tparam}
\alias{insert_tparam}
\title{Insert time-varying parameters into a tcovar matrix.}
\usage{
insert_tparam(tcovar, values, col_ind, tpar_inds)
}
\arguments{
\item{tcovar}{matrix into which the parameter values should be copied}

\item{values}{vector of values that should be copied in}

\item{col_ind}{C++ index for the column where the values should go}

\item{tpar_inds}{C++ indices for the vector elements that go in each row}
}
\value{
copy the values in place
}
\description{
Insert time-varying parameters into a tcovar matrix.
}
