% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interact}
\alias{interact}
\title{Interact character vectors to get a character vector of concatenated
combinations.}
\usage{
interact(...)
}
\arguments{
\item{...}{character vectors}
}
\value{
character vector whose elements are the concatenated elements of the
  input character vectors
}
\description{
Interact character vectors to get a character vector of concatenated
combinations.
}
\examples{
interact(sex=c("male", "female"), age=c("young", "adult", "old"))
}
