% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_progressive.R
\name{is_progressive}
\alias{is_progressive}
\title{Determines whether a model is progressive based on its flow matrix}
\usage{
is_progressive(flow_matrix)
}
\arguments{
\item{flow_matrix}{}
}
\value{
TRUE/FALSE indicating whether a model is progressive
}
\description{
Determines whether a model is progressive based on its flow matrix
}
