% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lna_density}
\alias{lna_density}
\title{Compute the density of an LNA path, reintegrating all LNA ODEs (as required,
e.g. after updating the parameters).}
\usage{
lna_density(path, lna_times, lna_pars, param_update_inds, flow_matrix,
  lna_pointer, set_pars_pointer)
}
\arguments{
\item{path}{list containing the lna_path, and the drift, residual, and
diffusion processes.}

\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{flow_matrix}{original flow matrix giving the changes to compartments
from each reaction}

\item{lna_pointer}{external pointer to LNA integration fcn}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
List containing the LNA path, the path of the residual
  process, the drift process, the residual process (conditional means), and
  diffusion process, along with the log density of the LNA path and the
  measurement process.
}
\description{
Compute the density of an LNA path, reintegrating all LNA ODEs (as required,
e.g. after updating the parameters).
}
