% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lna_incid2prev}
\alias{lna_incid2prev}
\title{Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale.}
\usage{
lna_incid2prev(path, flow_matrix, init_state, forcing_inds, forcing_matrix)
}
\arguments{
\item{path}{matrix containing the LNA path in terms of the counting
processes on transition events (incidence)}

\item{flow_matrix}{stoichiometry matrix (the transpose of the flow matrix)}

\item{init_state}{initial compartment counts on the natural scale}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.

The process can be re-expressed by left-multiplying each row in the path
matrix by the stoichiometry matrix: \eqn{X_t = X_0 + A'N_t}.}
}
\description{
Convert an LNA path from the counting process on transition events to the
compartment densities on their natural scale.
}
