% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lna_odes}
\alias{lna_odes}
\title{Construct a matrix containing the compartment counts at a sequence of census
times.}
\usage{
lna_odes(t, state, stemr_lnamod)
}
\arguments{
\item{t}{time}

\item{state}{vector containing the current values of the deterministic,
stochastic, and variance parts of the LNA}

\item{stemr_lnamod}{list containing the stemr LNA model objects, including
the external pointers to the hazard and jacobian functions (SEXPs), the
vectors of model parameters, constants, and time-varying covariates, and
the stoichiometry matrix (arma::mat).}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Construct a matrix containing the compartment counts at a sequence of census
times.
}

