% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lna_update.R
\name{lna_update}
\alias{lna_update}
\title{Sample a new LNA path via elliptical slice sampling.}
\usage{
lna_update(
  path,
  dat,
  params_cur,
  lna_ess_schedule,
  initdist_objects,
  tparam,
  pathmat_prop,
  censusmat,
  draws_prop,
  ess_draws_prop,
  emitmat,
  flow_matrix,
  stoich_matrix,
  times,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  param_inds,
  const_inds,
  tcovar_inds,
  initdist_inds,
  param_update_inds,
  census_indices,
  event_inds,
  measproc_indmat,
  svd_d,
  svd_U,
  svd_V,
  proc_pointer,
  set_pars_pointer,
  d_meas_pointer,
  do_prevalence,
  joint_initdist_update,
  return_ess_rec,
  step_size
)
}
\arguments{
\item{path}{path list}

\item{dat}{data matrix}

\item{params_cur}{matrix with current parameters}

\item{lna_ess_schedule}{}

\item{tparam}{list of time-varying parameters}

\item{pathmat_prop}{matrix for proposed paths}

\item{censusmat}{census matrix}

\item{emitmat}{emission matrix}

\item{flow_matrix}{flow matrix}

\item{stoich_matrix}{stoichiometry matrix}

\item{times}{vector of census times}

\item{forcing_inds}{forcing indices}

\item{forcing_tcov_inds}{time varying covariate forcings}

\item{forcings_out}{matrix with outflow from forcings}

\item{forcing_transfers}{transfer matrix}

\item{initdist_inds}{indices of initial compartment counts}

\item{census_indices}{indices where the path should be censused}

\item{event_inds}{event indices}

\item{measproc_indmat}{measurement process indices}

\item{svd_d}{SVD objects for LNA, NULL if using the ODE approx}

\item{svd_U}{SVD objects for LNA, NULL if using the ODE approx}

\item{svd_V}{SVD objects for LNA, NULL if using the ODE approx}

\item{proc_pointer}{C++ pointer for latent process}

\item{d_meas_pointer}{C++ pointer for emission distribution}

\item{do_prevalence}{should prevalence be computed}

\item{step_size}{initial step size for ODE solvers}
}
\value{
list with an updated LNA path along with its stochastic
  perturbations, the observed dat log-likelihood, and the lna
  log-likelihood, and a record of the number of elliptical slice sampling
  proposals
}
\description{
Sample a new LNA path via elliptical slice sampling.
}
