% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lna.R
\name{load_lna}
\alias{load_lna}
\title{Construct and compile the functions for proposing an LNA path, with
integration of the LNA ODEs accomplished using the Boost odeint library.}
\usage{
load_lna(lna_rates, messages)
}
\arguments{
\item{lna_rates}{list containing the LNA rate functions, derivatives, and
parameter codes}

\item{messages}{should messages be printed}
}
\value{
list containing the LNA pointers and calling code
}
\description{
Construct and compile the functions for proposing an LNA path, with
integration of the LNA ODEs accomplished using the Boost odeint library.
}

