% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lna_ess.R
\name{load_lna_ess}
\alias{load_lna_ess}
\title{Construct and compile the functions for proposing an LNA path within the
elliptical slice sampling algorithm, with integration of the LNA ODEs
accomplished using the Boost odeint library.}
\usage{
load_lna_ess(lna_rates, messages, atol, rtol)
}
\arguments{
\item{lna_rates}{list containing the LNA rate functions, derivatives, and
parameter codes}

\item{messages}{should messages be printed}
}
\value{
list containing the LNA pointers and calling code
}
\description{
Construct and compile the functions for proposing an LNA path within the
elliptical slice sampling algorithm, with integration of the LNA ODEs
accomplished using the Boost odeint library.
}
