% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ode.R
\name{load_ode}
\alias{load_ode}
\title{Construct and compile the functions for proposing an ODE path, with
integration of the deterministic mean ODEs accomplished using the Boost
odeint library.}
\usage{
load_ode(ode_rates, compile_ode, messages, atol, rtol, stepper)
}
\arguments{
\item{ode_rates}{list containing the ODE rate functions, derivatives, and
parameter codes}

\item{compile_ode}{if TRUE, code will be generated and compiled. If a
character string for the name of a file that does not yet exist, code will
be generated but not compiled. If the name of a file that exists in the
current working directory, the code in the file will be compiled.}

\item{messages}{should messages be printed}
}
\value{
list containing the ODE pointers and calling code
}
\description{
Construct and compile the functions for proposing an ODE path, with
integration of the deterministic mean ODEs accomplished using the Boost
odeint library.
}
