% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_pars_2_ode}
\alias{map_pars_2_ode}
\title{Map parameters to the deterministic mean incidence increments for a stochastic
epidemic model.}
\usage{
map_pars_2_ode(pathmat, ode_times, ode_pars, ode_param_inds, ode_tcovar_inds,
  init_start, param_update_inds, stoich_matrix, forcing_inds, forcing_matrix,
  step_size, ode_pointer, set_pars_pointer)
}
\arguments{
\item{pathmat}{matrix where the ODE path should be stored}

\item{ode_times}{vector of interval endpoint times}

\item{ode_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the ode_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
ode parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{ode_pointer}{external pointer to ode integration function.}

\item{set_pars_pointer}{external pointer to the function for setting the ode
parameters.}
}
\value{
List containing the ODE incidence and prevalence paths.
}
\description{
Map parameters to the deterministic mean incidence increments for a stochastic
epidemic model.
}
