% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_2_arr}
\alias{mat_2_arr}
\title{Copy a matrix into a slice of an array}
\usage{
mat_2_arr(dest, orig, ind)
}
\arguments{
\item{dest}{array into which to copy}

\item{orig}{matrix to copy}

\item{ind}{slice index (C++)}
}
\value{
copy a matrix into an array.
}
\description{
Copy a matrix into a slice of an array
}
