% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize_loglik_ode.R
\name{maximize_loglik_ode}
\alias{maximize_loglik_ode}
\title{Obtain estimates of stochastic epidemic model (SEM) parameters on their
estimation scales that maximize the observed data likelihood for a SEM
approximated by its deterministic ODE limit.}
\usage{
maximize_loglik_ode(stem_object, transformations = NULL, limits = NULL)
}
\arguments{
\item{stem_object}{stem_object for which the ODE code and measurement process
have been compiled.}

\item{transformations}{named list of functions for transforming to and from
the estimation scale, which should be unconstrained (i.e., R^n).}

\item{limits}{named list with vectors of upper and lower limits for
parameters ON THEIR ESTIMATION SCALE.}
}
\value{
list containing the parameter estimates on their estimation scales,
  Hessian of the data log likelihood evaluated at the maximum likelihood
  estimates on their estimation scales, the maximum likelihood ODE path, the
  data log-likelihood under the MLEs, the quasi-likelihood variance inflation
  factor, and an estimate of the sandwich standard errors for non-nuisance
  parameters.
}
\description{
NOTE: This should not be considered a reliable method for inference for SEMs.
Rather, it is, perhaps, useful as a method for obtaining initial values for
other methods.
}
