% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{mcmc_kernel}
\alias{mcmc_kernel}
\title{Specify an MCMC transition kernel}
\usage{
mcmc_kernel(
  parameter_blocks,
  lna_ess_control = NULL,
  initdist_ess_control = NULL,
  tparam_ess_control = NULL
)
}
\arguments{
\item{parameter_blocks}{list of parameter blocks with an MCMC kernel for each
block. Each list element should be generated by a call to the
\code{parblock} function}

\item{lna_ess_control}{list of control settings for updating LNA paths via
elliptical slice sampling, generated by a call to \code{lna_control}.}

\item{initdist_ess_control}{list of control settings for updating initial
distributions, generated by a call to \code{initdist_control}.}

\item{tparam_ess_control}{list of control settings for updating standard
normal variates for time-varying parameters via elliptical slice sampling.}
}
\value{
list containing the MCMC kernel.
}
\description{
Specify an MCMC transition kernel
}
\details{
Specifies the MCMC transition kernel for updating parameters, the
 latent epidemic process, time-varying parameters, and adaptive MCMC control.
 Parameter blocks group (time-homogeneous) parameters that are to be updated
 jointly. Each parameter block is updated via a multivariate
 Metropolis-Hastings update or a multivariate normal slice sampling update.
 Empirical covariances for these algorithms can be adapted using a global
 adaptive scaling algorithm (Andrieu and Thoms, 2008).

 References

 Andrieu, Christophe, and Johannes Thoms. "A tutorial on adaptive MCMC."
 Statistics and Computing 18.4 (2008): 343-373.
}
