% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_adaptive}
\alias{mvn_adaptive}
\title{Adaptive random walk Metropolis-Hastings transition kernel based on Roberts
and Rosenthal (2009).}
\usage{
mvn_adaptive(params_prop, params_cur, covmat, empirical_covmat, param_means,
  scaling, iteration, acceptances, nugget, nugget_weight, adapt_scale,
  adapt_shape, scale_start, shape_start, target, scale_cooling, max_scaling)
}
\arguments{
\item{params_prop}{vector in which the proposed parameters should be stored}

\item{params_cur}{vector containing the current parameter vector}

\item{covmat}{proposal covariance matrix}

\item{empirical_covmat}{empirical covariance matrix}

\item{param_means}{sample means of each of the parameters}

\item{scaling}{scale factor}

\item{iteration}{MCMC iteration number}

\item{acceptances}{number of acceptances}

\item{nugget}{adaptive RWMH nugget distribution}

\item{nugget_weight}{adaptive RWMH nugget contribution}

\item{scale_start}{iteration at which to begin scaling the supplied covariance matrix}

\item{shape_start}{iteration at which to begin estimating the empirical covariance matrix}

\item{target}{target acceptance rate when varying the scale}

\item{scale_cooling}{scale cooling rate}

\item{max_scaling}{maximum scale factor}
}
\value{
propose new parameter values in place and modify scaling and/or the empirical covariance matrix in place
}
\description{
Adaptive random walk Metropolis-Hastings transition kernel based on Roberts
and Rosenthal (2009).
}
