% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_rw}
\alias{mvn_rw}
\title{Random walk Metropolis-Hastings transition kernel.}
\usage{
mvn_rw(params_prop, params_cur, cov_chol)
}
\arguments{
\item{params_prop}{vector in which the proposed parameters should be stored}

\item{params_cur}{vector containing the current parameter vector}

\item{cov_chol}{upper triangular portion of the Cholesky decomposition of
the proposal covariance matrix}
}
\value{
propose new parameter values in place
}
\description{
Random walk Metropolis-Hastings transition kernel.
}
