% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_slice_sampler.R
\name{mvn_slice_sampler}
\alias{mvn_slice_sampler}
\title{Update model parameters via factor slice sampling}
\usage{
mvn_slice_sampler(model_params_est, model_params_nat, params_prop_est,
  params_prop_nat, mvn_direction, har_direction, mvnss_propvec, param_inds_Cpp,
  kernel_cov_chol, nugget, mvnss_bracket_width, n_expansions_mvnss,
  n_contractions_mvnss, path, pathmat_prop, data, priors, params_logprior_cur,
  lna_params_cur, lna_param_vec, tparam, censusmat, emitmat, flow_matrix,
  stoich_matrix, lna_times, forcing_inds, forcing_matrix, lna_param_inds,
  lna_const_inds, lna_tcovar_inds, lna_initdist_inds, param_update_inds,
  lna_event_inds, census_indices, measproc_indmat, svd_d, svd_U, svd_V,
  lna_pointer, lna_set_pars_pointer, d_meas_pointer, do_prevalence, step_size)
}
\arguments{
\item{model_params_est}{vector of model parameters on the estimation scale}

\item{model_params_nat}{vector of model parameters on the natural scale}

\item{params_prop_est}{vector for proposed model parameters on their
estimation scale}

\item{params_prop_nat}{vector for proposed parameters on their natural scale}

\item{mvn_direction}{vector for the hit and run direction}

\item{har_direction}{vector for the isotropic nugget}

\item{mvnss_propvec}{vector for the mvnss_component}

\item{param_inds_Cpp}{}

\item{kernel_cov_chol}{cholesky of the kernel covariance}

\item{nugget}{nugget variance, 0 if not adapting}

\item{mvnss_bracket_width}{width of the bracket}

\item{n_expansions_mvnss}{vector for number of mvn slice sampler expansions}

\item{n_contractions_mvnss}{vector for number of mvn slice sampler
contractions}

\item{path}{list containing the LNA path, N(0,1) draws, and likelihood}

\item{pathmat_prop}{matrix for the proposed LNA path}

\item{data}{matrix containing the data}

\item{priors}{list with functions for computing the prior density and
transformations to and from the estimation scale}

\item{params_logprior_cur}{log prior density of model parameters}

\item{lna_params_cur}{matrix with current LNA parameters, tcovar, tparam,
etc.}

\item{lna_param_vec}{vector for lna parameters}

\item{tparam}{list with time-varying parameters}

\item{censusmat}{matrix for storing the LNA path at census times}

\item{emitmat}{matrix for emission probabilities}

\item{flow_matrix}{flow matrix}

\item{stoich_matrix}{stoichiometry matrix}

\item{lna_times}{times at which the LNA is evaluated}

\item{forcing_inds}{indices at which forcings are applied}

\item{forcing_matrix}{matrix containing forcings}

\item{lna_param_inds}{indices for LNA parameters for computing emission probs}

\item{lna_const_inds}{indices for constants used in computing emission probs}

\item{lna_tcovar_inds}{indices for time-varying covariates}

\item{lna_initdist_inds}{index for where the initial compartment volumes
begin}

\item{param_update_inds}{indices for when LNA parameters should be updated}

\item{lna_event_inds}{codes for elementary events}

\item{census_indices}{indices for when the LNA path should be censused}

\item{measproc_indmat}{matrix giving the indices for data likelihood
contributions}

\item{svd_d}{vector for LNA singular values}

\item{svd_U}{matrix for LNA left singular vectors}

\item{svd_V}{matrix for LNA right singular vectors}

\item{lna_pointer}{external pointer for LNA}

\item{lna_set_pars_pointer}{external pointer for setting LNA parameters}

\item{d_meas_pointer}{external pointer for computing emission probabilities}

\item{do_prevalence}{should prevalence be computed}

\item{step_size}{initial step size for ODE stepper}
}
\value{
update the model parameters, path, and likelihood terms in place
}
\description{
Update model parameters via factor slice sampling
}
