% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_slice_sampler_ode.R
\name{mvn_slice_sampler_ode}
\alias{mvn_slice_sampler_ode}
\title{Update model parameters via factor slice sampling}
\usage{
mvn_slice_sampler_ode(model_params_est, model_params_nat, params_prop_est,
  params_prop_nat, mvn_direction, har_direction, mvnss_propvec,
  param_inds_Cpp, kernel_cov_chol, nugget, mvnss_bracket_width,
  n_expansions_mvnss, n_contractions_mvnss, path, pathmat_prop, data,
  priors, params_logprior_cur, ode_params_cur, ode_param_vec, tparam,
  censusmat, emitmat, flow_matrix, stoich_matrix, ode_times, forcing_inds,
  forcing_tcov_inds, forcings_out, forcing_transfers, ode_param_inds,
  ode_const_inds, ode_tcovar_inds, ode_initdist_inds, param_update_inds,
  ode_event_inds, census_indices, measproc_indmat, ode_pointer,
  ode_set_pars_pointer, d_meas_pointer, do_prevalence, step_size)
}
\arguments{
\item{model_params_est}{vector of model parameters on the estimation scale}

\item{model_params_nat}{vector of model parameters on the natural scale}

\item{params_prop_est}{vector for proposed model parameters on their
estimation scale}

\item{params_prop_nat}{vector for proposed parameters on their natural scale}

\item{mvn_direction}{}

\item{har_direction}{vector for the hit and run direction}

\item{mvnss_propvec}{}

\item{param_inds_Cpp}{}

\item{kernel_cov_chol}{}

\item{nugget}{}

\item{mvnss_bracket_width}{}

\item{n_expansions_mvnss}{}

\item{n_contractions_mvnss}{}

\item{path}{list containing the ode path and likelihood}

\item{pathmat_prop}{}

\item{data}{matrix containing the data}

\item{priors}{list with functions for computing the prior density and
transformations to and from the estimation scale}

\item{params_logprior_cur}{log prior density of model parameters}

\item{ode_params_cur}{matrix with current ode parameters, tcovar, tparam,
etc.}

\item{ode_param_vec}{vector for ode parameters}

\item{tparam}{list with time-varying parameters}

\item{censusmat}{matrix for storing the ode path at census times}

\item{emitmat}{matrix for emission probabilities}

\item{flow_matrix}{flow matrix}

\item{stoich_matrix}{stoichiometry matrix}

\item{ode_times}{times at which the ode is evaluated}

\item{forcing_inds}{indices at which forcings are applied}

\item{ode_param_inds}{indices for ode parameters for computing emission probs}

\item{ode_const_inds}{indices for constants used in computing emission probs}

\item{ode_tcovar_inds}{indices for time-varying covariates}

\item{ode_initdist_inds}{index for where the initial compartment volumes
begin}

\item{param_update_inds}{indices for when ode parameters should be updated}

\item{ode_event_inds}{codes for elementary events}

\item{census_indices}{indices for when the ode path should be censused}

\item{measproc_indmat}{matrix giving the indices for data likelihood
contributions}

\item{ode_pointer}{external pointer for ode}

\item{ode_set_pars_pointer}{external pointer for setting ode parameters}

\item{d_meas_pointer}{external pointer for computing emission probabilities}

\item{do_prevalence}{should prevalence be computed}

\item{step_size}{initial step size for ODE stepper}

\item{forcing_matrix}{matrix containing forcings}
}
\value{
update the model parameters, path, and likelihood terms in place
}
\description{
Update model parameters via factor slice sampling
}
