% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnmh_update.R
\name{mvnmh_update}
\alias{mvnmh_update}
\title{Multivariate normal Metropolis-Hastings update}
\usage{
mvnmh_update(
  param_blocks,
  ind,
  iter,
  parmat,
  dat,
  path,
  pathmat_prop,
  tparam,
  census_times,
  flow_matrix,
  stoich_matrix,
  censusmat,
  emitmat,
  param_vec,
  param_inds,
  const_inds,
  tcovar_inds,
  initdist_inds,
  param_update_inds,
  census_indices,
  event_inds,
  measproc_indmat,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  proc_pointer,
  set_pars_pointer,
  d_meas_pointer,
  do_prevalence,
  step_size,
  svd_d = NULL,
  svd_U = NULL,
  svd_V = NULL
)
}
\arguments{
\item{param_blocks}{list of parameter blocks}

\item{ind}{index of parameter block to be updated}

\item{iter}{MCMC iteration}

\item{dat}{data matrix}

\item{path}{path list}

\item{pathmat_prop}{matrix for proposed paths}

\item{tparam}{list of time-varying parameters}

\item{flow_matrix}{flow matrix}

\item{stoich_matrix}{stoichiometry matrix}

\item{censusmat}{census matrix}

\item{emitmat}{emission matrix}

\item{param_vec}{parameter vector for use in emission distribution}

\item{initdist_inds}{indices of initial compartment counts}

\item{census_indices}{indices where the path should be censused}

\item{event_inds}{event indices}

\item{measproc_indmat}{measurement process indices}

\item{forcing_inds}{forcing indices}

\item{forcing_tcov_inds}{time varying covariate forcings}

\item{forcings_out}{matrix with outflow from forcings}

\item{forcing_transfers}{transfer matrix}

\item{proc_pointer}{C++ pointer for latent process}

\item{d_meas_pointer}{C++ pointer for emission distribution}

\item{do_prevalence}{should prevalence be computed}

\item{step_size}{initial step size for ODE solvers}

\item{svd_d, svd_U, svd_V}{SVD objects for LNA, NULL if using the ODE approx}

\item{params_cur}{matrix with current parameters}

\item{params_prop}{matrix with proposed parameters}

\item{times}{vector of census times}
}
\value{
update the model parameters, path, and likelihood
}
\description{
Multivariate normal Metropolis-Hastings update
}
