% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnss_control.R
\name{mvnss_control}
\alias{mvnss_control}
\title{Generate a list of settings for multivariate normal slice sampling}
\usage{
mvnss_control(
  n_updates = 1,
  bracket_width = 1,
  cov_update_interval = 1,
  bracket_limits = c(0, Inf),
  scale_constant = 0.5,
  scale_cooling = 0.5 + 1e-05,
  step_size = 1,
  stop_adaptation = 0,
  adaptation_offset = 0,
  nugget = NULL,
  nugget_cooling = 0.9,
  nugget_step_size = NULL
)
}
\arguments{
\item{n_updates}{number of updates per iteration.}

\item{cov_update_interval}{how often the empirical covariance matrix should
be updated}

\item{bracket_limits}{limits for the slice bracket width, defaults to
(0,Inf).}

\item{scale_constant}{constant multiple of the adaptations determined by
\code{scale_cooling}.}

\item{scale_cooling}{rate at which to cool the adaptation, defaults to 0.5.
Adaptation contributions are governed by a harmonic sequence:
scale_constant/(iteration/step_size+1)^scale_cooling. The
\code{plot_adaptations} function may be used to plot the adaptation factors.}

\item{step_size}{adaptation increment for each iteration, defaults to 1.}

\item{stop_adaptation}{iteration at which adaptation should be terminated, 
defaults to 0 for no adaptation.}

\item{adaptation_offset}{iteration offset}

\item{nugget}{nugget for proposal covariance}

\item{nugget_cooling}{rate at which to cool the nugget, defaults to 0.9.
Adaptation contributions are governed by a harmonic sequence:
scale_constant/(iteration/step_size+1)^scale_cooling. The
\code{plot_adaptations} function may be used to plot the adaptation
factors.}

\item{nugget_step_size}{increment for each iteration, defaults to 100/number
of iterations.}

\item{initial_bracket_width}{initial width of the slice bracket}
}
\value{
list with control settings for multivariate normal slice sampling
}
\description{
Generate a list of settings for multivariate normal slice sampling
}
