% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnss_settings.R
\name{mvnss_settings}
\alias{mvnss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
mvnss_settings(n_mvnss_updates = 1, initial_bracket_width = 1,
  cov_update_interval = 1)
}
\arguments{
\item{n_mvnss_updates}{number of non-isotropoc hit-and-run updates.}

\item{cov_update_interval}{how often the empirical covariance matrix should be updated}
}
\value{
list with additional settings for hit and run slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
