% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrn}
\alias{mvrn}
\title{Produce samples from a multivariate normal density using the singular value
decomposition as in the mvtnorm R package.}
\usage{
mvrn(n, mu, sigma)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}

\item{sigma}{covariance matrix}
}
\description{
Produce samples from a multivariate normal density using the singular value
decomposition as in the mvtnorm R package.
}

