% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalise}
\alias{normalise}
\title{normalise a vector in place}
\usage{
normalise(v, p)
}
\arguments{
\item{v}{vector to be normalised}

\item{p}{norm}
}
\value{
normalise vector in place
}
\description{
normalise a vector in place
}
