% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalise2}
\alias{normalise2}
\title{return a normalised vector}
\usage{
normalise2(v, p)
}
\arguments{
\item{v}{vector to be normalised}

\item{p}{norm}
}
\value{
normalised vector
}
\description{
return a normalised vector
}
