% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parblock.R
\name{parblock}
\alias{parblock}
\title{Define a parameter block for an MCMC kernel}
\usage{
parblock(pars, alg, sigma, control = NULL)
}
\arguments{
\item{pars}{character vector of parameter names on their estimation scale
(i.e., corresponding to row/column names in the kernel covariance matrix)}

\item{alg}{either "mvnmh"  or "mvnss" for multivariate normal metropolis
hastings updates or multivariate normal slice sampling updates,
respectively.}

\item{sigma}{initial covariance matrix for the parameter block, possibly to
be adapted.}

\item{control}{list of mcmc control settings, generated by a call to
\code{mvnmh_control} or \code{mvnss_control} as appropriate.}
}
\value{
parameter block for use in MCMC kernel
}
\description{
Define a parameter block for an MCMC kernel
}
