% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pars2lnapars2}
\alias{pars2lnapars2}
\title{Insert parameters into the first row of a parameter matrix}
\usage{
pars2lnapars2(lnapars, parameters, c_start)
}
\arguments{
\item{lnapars}{matrix of lna/ode parameters, constants, and time-varying covars}

\item{parameters}{vector of parameters to be copied into the matrix}

\item{c_start}{index of the initial column}
}
\value{
modifies the lna parameter matrix in place
}
\description{
Insert parameters into the first row of a parameter matrix
}
