% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pars2parmat}
\alias{pars2parmat}
\title{Insert parameters into the first row of a parameter matrix}
\usage{
pars2parmat(parmat, pars, colinds, rowinds = 0L)
}
\arguments{
\item{parmat}{parameter matrics}

\item{pars}{vector of parameters to insert}

\item{colinds}{vector of column indices}

\item{rowinds}{vector of row indices, just the first row by default.}
}
\value{
modifies the parameter matrix in place
}
\description{
Insert parameters into the first row of a parameter matrix
}
