% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lna.R
\name{parse_lna}
\alias{parse_lna}
\title{Instatiate the C++ rate functions for the LNA.}
\usage{
parse_lna(rates, hazard_fcns, rate_derivs, messages = TRUE)
}
\arguments{
\item{rates}{list of rate functions}

\item{rate_derivs}{character vector of partial derivatives of the rate
functions.}

\item{messages}{logical; print a message that the rates are being compiled}
}
\value{
Vector of strings that serve as function pointers.
}
\description{
Instatiate the C++ rate functions for the LNA.
}

