% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lna_rates.R
\name{parse_lna_rates}
\alias{parse_lna_rates}
\title{Parse the LNA rates so they can be compiled.}
\usage{
parse_lna_rates(
  lna_rates,
  param_codes,
  const_codes,
  tcovar_codes,
  lna_comp_codes
)
}
\arguments{
\item{lna_rates}{character vector of LNA rates}

\item{param_codes}{named numeric vector of parameter codes}

\item{const_codes}{named numeric vector of constant codes}

\item{tcovar_codes}{named numeric vector of time-varying covariate codes}

\item{lna_comp_codes}{named numeric vector of LNA compartment codes}
}
\value{
string snippets for the LNA that can be compiled
}
\description{
Parse the LNA rates so they can be compiled.
}
