% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_meas_procs.R
\name{parse_meas_procs}
\alias{parse_meas_procs}
\title{Instatiate the C++ emission probability functions for simulation and density
evaluation for a stochastic epidemic model measurement process and return a
vector of function pointers.}
\usage{
parse_meas_procs(meas_procs, messages = TRUE)
}
\arguments{
\item{meas_procs}{list of measurement process functions}

\item{messages}{logical; print a message that the rates are being compiled?}
}
\value{
Two vector of strings that serve as function pointers.
}
\description{
Instatiate the C++ emission probability functions for simulation and density
evaluation for a stochastic epidemic model measurement process and return a
vector of function pointers.
}

