% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ode_fcns.R
\name{parse_ode_fcns}
\alias{parse_ode_fcns}
\title{Compile the ODE functions for a stochastic epidemic model}
\usage{
parse_ode_fcns(ode_rates, flow_matrix, messages)
}
\arguments{
\item{ode_rates}{character vector of hazard rates, as outputted by
\code{\link{build_ode_rates}}}

\item{flow_matrix}{matrix indicating the changes to each compartment per
event}

\item{messages}{logical; print a message that the rates are being compiled}
}
\value{
external function pointers for integrating the ODE
}
\description{
Compile the ODE functions for a stochastic epidemic model
}

