% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ode_rates.R
\name{parse_ode_rates}
\alias{parse_ode_rates}
\title{Parse the ODE rates so they can be compiled.}
\usage{
parse_ode_rates(ode_rates, param_codes, const_codes, tcovar_codes,
  ode_comp_codes)
}
\arguments{
\item{ode_rates}{character vector of ODE rates}

\item{param_codes}{named numeric vector of parameter codes}

\item{const_codes}{named numeric vector of constant codes}

\item{tcovar_codes}{named numeric vector of time-varying covariate codes}

\item{ode_comp_codes}{named numeric vector of ODE compartment codes}
}
\value{
string snippets for the ODE that can be compiled
}
\description{
Parse the ODE rates so they can be compiled.
}
