% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_parameter_blocks.R
\name{parse_parameter_blocks}
\alias{parse_parameter_blocks}
\title{Generate a list of objects used in block updating MCMC parameters}
\usage{
parse_parameter_blocks(param_blocks, param_names_est)
}
\arguments{
\item{param_names_est}{character vector of (estimation scale) parameter names}

\item{parameter_blocks}{list of numeric or character vectors identifying
parameters in each block}
}
\value{
list containing objects used in block updates to MCMC parameters
}
\description{
Generate a list of objects used in block updating MCMC parameters
}
