% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rate.R
\name{parse_rate}
\alias{parse_rate}
\title{Transform a text snippet for a rate function into C++ text for the rate
function.}
\usage{
parse_rate(rate, parameters, compartment_names, constants, strata, adjacency)
}
\arguments{
\item{rate}{string snippet declaring how to compute the rate, must be a
function of parameters, compartment_strata, and constants, and must be
valid C++ code.}

\item{parameters}{vector of parameters}

\item{compartment_names}{vector of compartment names}

\item{constants}{vector of constant names}

\item{strata}{character vector of strata names to which the rate applies, or
numeric vector of strata numbers}

\item{adjacency}{adjacency matrix}
}
\value{
string that can be evaluated and compiled to create a C++ function
  for computing the rate.
}
\description{
Transform a text snippet for a rate function into C++ text for the rate
function.
}

