% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rates_exact.R
\name{parse_rates_exact}
\alias{parse_rates_exact}
\title{Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.}
\usage{
parse_rates_exact(rates, compile_rates, messages = TRUE)
}
\arguments{
\item{rates}{list of rate functions}

\item{compile_rates}{if TRUE, code will be generated and compiled. If a
character string for the name of a file that does not yet exist, code will
be generated but not compiled. If the name of a file that exists in the
current working directory, the code in the file will be compiled.}

\item{messages}{logical; print a message that the rates are being compiled}
}
\value{
Two vector of strings that serve as function pointers.
}
\description{
Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.
}
