% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pca_adaptive}
\alias{pca_adaptive}
\title{Adaptive principal component metropolis proposal.}
\usage{
pca_adaptive(params_prop, params_cur, eigenvalues, eigenvectors,
  proposal_scaling, direction)
}
\arguments{
\item{params_prop}{vector in which the proposed parameters should be stored}

\item{params_cur}{vector containing the current parameter vector}

\item{eigenvalues}{vector of eigenvalues of the proposal covariance matrix}

\item{eigenvectors}{matrix of eigenvectors for the proposal covariance}

\item{proposal_scaling}{vector of component scaling parameters}

\item{direction}{component to be used in the proposal}
}
\value{
propose new parameter values in place
}
\description{
Adaptive principal component metropolis proposal.
}
