% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_settings.R
\name{pcm_settings}
\alias{pcm_settings}
\title{Specify settings for principal component metropolis}
\usage{
pcm_settings(weight_update_interval = NULL, pcm_update_schedule = "random",
  n_pcm_updates = 1)
}
\arguments{
\item{weight_update_interval}{number of iterations between updates to the
direction sampling distribution if using adaptive principal components
Metropolis. Defaults to 100 times the number of parameters if not specified.
The weights are the sqrt(n)^th root of the eigenvalues, where n is the number of
model parameters, normalized to sum to 1.}

\item{pcm_update_schedule}{either "random" (default) or "sequential". If
"random", each iteration consists of one parameter update in a direction
chosen at random according to the weight distribution. If "sequential", each
MCMC iteration consists of an update in each direction.}

\item{n_pcm_updates}{If the principal component metropolis schedule is random,
the number of directions, chosen at random, that should be updated per
iteration.}
}
\value{
list containing settings for principal component metropolis
}
\description{
Specify settings for principal component metropolis
}
