% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTopTags.R
\name{plotTopTags}
\alias{plotTopTags}
\title{Function plots top genes}
\usage{
plotTopTags(counts, groups)
}
\arguments{
\item{counts}{matrix (small) for top genes.}

\item{groups}{groups for following collumns in counts.}
}
\description{
Function plots top genes
}

