% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_initdist_objects.R
\name{prepare_initdist_objects}
\alias{prepare_initdist_objects}
\title{Prepare initial distribution list for MCMC}
\usage{
prepare_initdist_objects(initializer, param_codes, comp_size_vec)
}
\arguments{
\item{initializer}{initializer list}

\item{param_codes}{vector of parameter codes in the params matrix}

\item{comp_size_vec}{vector of compartment volumes}
}
\value{
list of objects for MCMC
}
\description{
Prepare initial distribution list for MCMC
}
