% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lna_ess_schedule.R
\name{prepare_lna_ess_schedule}
\alias{prepare_lna_ess_schedule}
\title{Prepare an LNA elliptical slice sampling schedule}
\usage{
prepare_lna_ess_schedule(stem_object, initializer, lna_ess_control)
}
\arguments{
\item{stem_object}{stem_object list}

\item{initializer}{initializer list}

\item{lna_ess_control}{LNA control list}
}
\value{
LNA ESS schedule
}
\description{
Prepare an LNA elliptical slice sampling schedule
}
