% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Specify the prior distribution for a model parameter}
\usage{
prior(parameter, distribution, hyperpars, scale = "linear")
}
\arguments{
\item{parameter}{character string giving the name of the model parameter}

\item{distribution}{character string giving the prior distribution for the
parameter, may be specified according to the table below}

\item{hyperpars}{named vector of prior distribution hyperparameters with
names matching the arguments of the appropriate distribution in the table
below}

\item{scale}{on which the prior is specified. Defaults to "linear", may be
specified as "log" or "logit". The appropriate transformation will be
applied automatically for the proposals.}
}
\value{
list to be used in evaluating the prior densities of model
  parameters.
}
\description{
Specify the prior distribution for a model parameter
}
\section{Specification of prior distributions}{


  The distributions and hyperparameters should be specified according to the
  table below

  \tabular{lll}{ distribution \tab abbreviation \tab hyperparameters \cr Beta
  \tab beta \tab shape1, shape2, ncp \cr Cauchy \tab cauchy \tab location,
  scale \cr Gamma \tab gamma \tab shape, rate = 1, scale = 1/rate \cr Normal
  \tab norm \tab mean, sd \cr Truncated normal (a,b) \tab tnorm \tab mean,
  sd, a, b \cr Student t \tab t \tab df, ncp \cr Uniform \tab unif \tab min,
  max}

  N.B. The prior for t0 should be specified through the
  \code{\link{t0_prior}} function.
}

\examples{
prior("theta", "norm", c(mean = 1, sd = 2), scale = "log")
}
