% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna_approx}
\alias{propose_lna_approx}
\title{Simulate an approximate LNA path using a non-centered parameterization for the
log-transformed counting process LNA. Resample the initial path in place, then
update with elliptical slice sampling.}
\usage{
propose_lna_approx(lna_times, lna_pars, lna_param_inds, lna_tcovar_inds,
  init_start, param_update_inds, stoich_matrix, forcing_inds, forcing_matrix,
  max_attempts, nsim, ess_updates, ess_warmup, step_size, lna_pointer,
  set_pars_pointer)
}
\arguments{
\item{lna_times}{vector of interval endpoint times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{max_attempts}{maximum number of tries to restart negative increments if 
reject_negatives is false.}

\item{nsim}{number of paths to simulate}

\item{ess_updates}{number of elliptical slice sampling updates between samples.}

\item{ess_warmup}{number of elliptical slice sampling updates before the first
sample.}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{lna_pointer}{external pointer to the compiled LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting LNA pars.}
}
\value{
list containing the stochastic perturbations (i.i.d. N(0,1) draws) and
the LNA path on its natural scale which is determined by the perturbations.
}
\description{
Simulate an approximate LNA path using a non-centered parameterization for the
log-transformed counting process LNA. Resample the initial path in place, then
update with elliptical slice sampling.
}
