% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna_ess}
\alias{propose_lna_ess}
\title{Simulate a data matrix from the measurement process of a stochastic epidemic
model.}
\usage{
propose_lna_ess(path_cur, lna_times, lna_pars, param_update_inds, flow_matrix,
  lna_pointer_ess, set_ess_pars_pointer)
}
\arguments{
\item{path_cur}{list containing the LNA objects for the current path}

\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{flow_matrix}{original flow matrix giving the changes to compartments
from each reaction}

\item{init_state}{rowvec giving the initial compartment counts}

\item{log_scale}{is the LNA applied on the log scale?}

\item{lna_pointer}{external pointer to LNA integration fcn}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
array with slices containing the LNA path, the path of the residual
  process, the drift process, the residual process (conditional means), and
  diffusion process.
}
\description{
Simulate a data matrix from the measurement process of a stochastic epidemic
model.
}

