% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_lna_path.R
\name{propose_lna_path}
\alias{propose_lna_path}
\title{Propose a latent path using the linear noise approximation and return the
paths of the LNA ODEs.}
\usage{
propose_lna_path(parameters, stoich_matrix, lna_pointer, times,
  initdist_parameters, fixed_inits, param_update_inds, drift_inds, resid_inds,
  diff_inds, log_scale, incidence_codes = NULL)
}
\arguments{
\item{parameters}{list of lna parameters, the lna function pointer, and the
stoichiometry matrix}

\item{stoich_matrix}{the stoichiometry matrix, i.e. the transpose of the flow
matrix.}

\item{lna_pointer}{external pointer to the LNA function}

\item{times}{vector of times at which the LNA is to be evaluated}

\item{initdist_parameters}{vector of parameters for the initial state}

\item{fixed_inits}{is the initial state fixed or random?}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{drift_inds}{vector of indices corresponding to the drift process within
the vector holding the ODEs}

\item{resid_inds}{vector of indices corresponding to the residual process
within the vector holding the ODEs}

\item{diff_inds}{vector of indices corresponding to the diffusion process
within the vector holding the ODEs}

\item{log_scale}{logical}

\item{incidence_codes}{vector of incidence codes}

\item{add_noise}{add some noise to the diagonal of the diffusion matrix,
defaults to FALSE}
}
\value{
list with the proposed LNA path, along with the drift, residual, and
  diffusion processes
}
\description{
Propose a latent path using the linear noise approximation and return the
paths of the LNA ODEs.
}

