% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_lna_path_ess.R
\name{propose_lna_path_ess}
\alias{propose_lna_path_ess}
\title{Propose a latent LNA path with zero mean for use in the elliptical slice
sampler and return the paths of the LNA ODEs.}
\usage{
propose_lna_path_ess(path_cur, parameters, stoich_matrix, lna_ess_pointer,
  times, initdist_parameters, fixed_inits, param_update_inds, drift_inds,
  resid_inds, log_scale, incidence_codes = NULL)
}
\arguments{
\item{path_cur}{list containing the current path and LNA ODEs}

\item{parameters}{list of lna parameters, the lna function pointer, and the
stoichiometry matrix}

\item{stoich_matrix}{the stoichiometry matrix, i.e. the transpose of the flow
matrix.}

\item{times}{vector of times at which the LNA is to be evaluated}

\item{initdist_parameters}{vector of initial state parameters}

\item{fixed_inits}{is the initial state fixed or random?}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{drift_inds}{vector of indices corresponding to the drift process within
the vector holding the ODEs}

\item{resid_inds}{vector of indices corresponding to the residual process
within the vector holding the ODEs}

\item{log_scale}{logical}

\item{incidence_codes}{vector of incidence codes}

\item{fixed_inits}{are the initial states fixed?}

\item{initdist_parameters}{mean vector for the initial compartment counts}

\item{diff_inds}{vector of indices corresponding to the diffusion process
within the vector holding the ODEs}

\item{lna_pointer}{external pointer to the LNA function}
}
\value{
list with the proposed LNA path, along with the drift, residual, and
  diffusion processes
}
\description{
Propose a latent LNA path with zero mean for use in the elliptical slice
sampler and return the paths of the LNA ODEs.
}

