% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna_t}
\alias{propose_lna_t}
\title{Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.}
\usage{
propose_lna_t(lna_times, lna_pars, init_start, param_update_inds, stoich_matrix,
  forcing_inds, forcing_matrix, step_size, df, lna_pointer, set_pars_pointer)
}
\arguments{
\item{lna_times}{vector of interval endpoint times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{lna_pointer}{external pointer to LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
list containing the stochastic perturbations (i.i.d. N(0,1) draws) and
the LNA path on its natural scale which is determined by the perturbations.
}
\description{
Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.
}
