% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_fcns_4_ode.R
\name{rate_fcns_4_ode}
\alias{rate_fcns_4_ode}
\title{Convert unparsed rate functions into deterministic mean ODE functions.}
\usage{
rate_fcns_4_ode(rate_fcns, compartment_codes, flow_matrix)
}
\arguments{
\item{rate_fcns}{list of rates, as returned by stem_dynamics}

\item{compartment_codes}{vector of compartment codes}

\item{flow_matrix}{matrix indicating compartment flow, whose transpose is the
stoichiometry matrix}
}
\value{
list containing a character vector with the rate functions and a
  named numeric vector of counting process compartment codes
}
\description{
Convert unparsed rate functions into deterministic mean ODE functions.
}
