% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rate_update_event}
\alias{rate_update_event}
\title{Identify which rates to update when a state transition event occurs.}
\usage{
rate_update_event(rate_inds, M, event_code)
}
\arguments{
\item{rate_inds}{vector of rate indices to be modified}

\item{M}{adjacency matrix for which rates need to be updated in response to a transition}

\item{event_code}{column in the rate adjacency matrix}
}
\value{
modifies logical vector stating which rates need to be updated
}
\description{
Identify which rates to update when a state transition event occurs.
}

