% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rate_update_tcovar}
\alias{rate_update_tcovar}
\title{Identify which rates to update based on changes in the time-varying covariates.}
\usage{
rate_update_tcovar(rate_inds, M, I)
}
\arguments{
\item{rate_inds}{vector of rate indices to be modified}

\item{M}{time-varying covariate adjacency matrix}

\item{I}{logical vector indicating which covariates changed at a particular time.}
}
\value{
logical vector stating which rates need to be updated
}
\description{
Identify which rates to update based on changes in the time-varying covariates.
}
