% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_nugget.R
\name{reset_nugget}
\alias{reset_nugget}
\title{Reset the nugget based based on a sample}
\usage{
reset_nugget(nugget, parameter_samples_est, start, end, lna_initdist_inds)
}
\arguments{
\item{nugget}{current value, a scalar}

\item{start}{beginning of range to include}

\item{end}{end of range to include}

\item{lna_initdist_inds}{indices for which parameters are initial distribution parameters}

\item{parameter_samples}{parameter samples on their estimation}
}
\value{
vector of nugget standard deviations
}
\description{
Reset the nugget based based on a sample
}
