% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reset_slice_ratios}
\alias{reset_slice_ratios}
\title{Reset counters for interval expansions/contractions and slice ratios}
\usage{
reset_slice_ratios(
  n_expansions,
  n_contractions,
  n_expansions_c,
  n_contractions_c,
  slice_ratios
)
}
\arguments{
\item{n_expansions}{vector with number of expansion}

\item{n_contractions}{vector with number of contractions}

\item{n_expansions_c}{cumulative numbers of expansions}

\item{n_contractions_c}{cumulative numbers of contractions}

\item{slice_ratios}{vector for storing ratio of cumulative number of 
expansions over number of interval width changes}
}
\value{
reset objects in place
}
\description{
Reset counters for interval expansions/contractions and slice ratios
}
