% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reset_vec}
\alias{reset_vec}
\title{Reset a vector by filling it with zeros}
\usage{
reset_vec(v)
}
\arguments{
\item{v}{vector to fill with zeros}
}
\value{
reset vector in place
}
\description{
Reset a vector by filling it with zeros
}
