% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_census_path}
\alias{retrieve_census_path}
\title{Insert the compartment counts at a sequence of census times into an existing census matrix.}
\usage{
retrieve_census_path(censusmat, path, census_times, census_columns)
}
\arguments{
\item{censusmat}{matrix of compartment counts at census times, to be updated}

\item{path}{matrix containing the path to be censused.}

\item{census_times}{vector of census times.}

\item{census_columns}{vector of column indices to be censused (C++ indexing
beginning at 0).}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Insert the compartment counts at a sequence of census times into an existing census matrix.
}

