% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvtn}
\alias{rmvtn}
\title{Produce samples from a multivariate normal density using the Cholesky
decomposition}
\usage{
rmvtn(n, mu, sigma)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}

\item{sigma}{covariance matrix

(source: http://gallery.rcpp.org/articles/simulate-multivariate-normal/)}
}
\description{
Produce samples from a multivariate normal density using the Cholesky
decomposition
}
