% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsbln.R
\name{rsbln}
\alias{rsbln}
\title{Stick-Breaking Logit-Normal}
\usage{
rsbln(n, stick_means, stick_sds, stick_size)
}
\arguments{
\item{n}{number of observations}

\item{stick_means}{vector of stick means on the logit scale}

\item{stick_sds}{vector of stick sds on the logit scale}

\item{stick_size}{size by which to scale the observations (population size)}
}
\value{
samples from a stick breaking process
}
\description{
Stick-Breaking Logit-Normal
}
