% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_unit_sphere}
\alias{sample_unit_sphere}
\title{Sample the unit sphere.}
\usage{
sample_unit_sphere(v)
}
\arguments{
\item{v}{vector to fill with a vector of draws on the unit sphere}
}
\value{
draw new values in place
}
\description{
Sample the unit sphere.
}
