% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ess_rec.R
\name{save_ess_rec}
\alias{save_ess_rec}
\title{Save the elliptical slice sampling record}
\usage{
save_ess_rec(
  ess_record,
  ess_rec_ind,
  lna_ess_schedule = NULL,
  tparam = NULL,
  initdist_ess_control = NULL
)
}
\arguments{
\item{ess_record}{list with objects for recording the ESS record}

\item{ess_rec_ind}{C++ record index}

\item{lna_ess_schedule}{list with lna ESS objects}

\item{tparam}{list with time varying parameters}

\item{initdist_objects}{list with initial distribution objects}
}
\value{
inserts ESS steps and angles into the ESS record objects
}
\description{
Save the elliptical slice sampling record
}
