% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ess_record.R
\name{save_ess_record}
\alias{save_ess_record}
\title{Record ESS steps and angles}
\usage{
save_ess_record(
  ess_record,
  ess_rec_ind,
  lna_ess_schedule = NULL,
  initdist_objects = NULL,
  tparam = NULL
)
}
\arguments{
\item{ess_record}{object where the steps and angles are to be stored}

\item{ess_rec_ind}{index}

\item{lna_ess_schedule}{lna ESS list}

\item{initdist_objects}{}

\item{tparam}{tparam ESS list}
}
\value{
save the ESS record
}
\description{
Record ESS steps and angles
}
