% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mcmc_sample.R
\name{save_mcmc_sample}
\alias{save_mcmc_sample}
\title{Save an MCMC sample}
\usage{
save_mcmc_sample(
  mcmc_samples,
  rec_ind,
  path,
  parmat,
  param_blocks,
  initdist_objects,
  tparam,
  tparam_inds,
  method
)
}
\arguments{
\item{mcmc_samples}{list with objects for recording MCMC samples}

\item{rec_ind}{C++ record index}

\item{path}{latent path}

\item{param_blocks}{list of parameter blocks}

\item{initdist_objects}{list of initial distribution objects}

\item{tparam}{list with time-varying parameters}

\item{method}{method for approximating the latent epidemic, either "lna" or
"ode"}
}
\value{
inserts all parameters into their respective MCMC record
}
\description{
Save an MCMC sample
}
