% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsbln.R
\name{sbln_normal_to_volume}
\alias{sbln_normal_to_volume}
\title{Convert Normal Draws to Stick-Breaking Logit-Normal Draws}
\usage{
sbln_normal_to_volume(normal_draws, stick_means, stick_sds, stick_size)
}
\arguments{
\item{normal_draws}{a matrix of normal draws to convert to compartment volumes}

\item{stick_means}{vector of stick means on the logit scale}

\item{stick_sds}{vector of stick sds on the logit scale}

\item{stick_size}{size to scale the samples by}
}
\value{
samples from a stick breaking process
}
\description{
Convert Normal Draws to Stick-Breaking Logit-Normal Draws
}
