% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality}
\alias{seasonality}
\title{Produces a string for seasonal terms to be added to a rate function, supplied
as the last argument in a rate list.}
\usage{
seasonality(period = 52, s_params = NULL, log = TRUE,
  common_seasonality = TRUE)
}
\arguments{
\item{period}{period of the season, defaults to 52 for weeks. May also be
supplied as a vector if S>1 in order to specify multiple terms with
different periods.}

\item{log}{should the seasonal covariates enter into the unlumped rates on
the log scale? If true (default), each rate, \eqn{\lambda}, is log-linear
in the seasonal terms. For example, if \eqn{\lambda_{S,I} = (\beta I +
\alpha(t))S}, then \deqn{\alpha(t) = exp(\sum \gamma_p \times sin(2\pi
p\times t) + \delta_p \times cos(2\pi p\times t))}}

\item{s_param_values}{seasonality parameter values (or initial values), a
2*length(period) numeric vector, with all sine parameters, then all cosine
parameters, and supplied in the same order as period. If no values are
supplied, the parameter values default to 0. Thus, if \code{period = c(12,
52)}, the parameter values would be, for example, \code{c(SIN_12 = 1,
SIN_52 = 2, COS_12 = 3, COS_52 = 4)}.}

\item{common_seasonlity}{logical for whether the seasonal terms should be
shared among strata (TRUE) or whether each stratum should have its own
seasonality}
}
\value{
string formula for the seasonal terms to be added to a rate.
}
\description{
Supplies a string for the seasonal terms \deqn{\sum \gamma_p
  \times sin(2\pi p\times t) + \delta_p \times cos(2\pi p\times t)} to be
  added to a rate function. The parameters \eqn{\gamma period} and
  \eqn{\delta period} will automatically be named to reflect whether the
  seasonal terms are common to all strata, and then appended to the parameter
  vector. If \code{log=TRUE}, the seasonal terms are exponentiated. Thus,
  \deqn{\sum \gamma_p \times sin(2\pi p\times t) + \delta_p \times cos(2\pi
  p\times t)}
}
\examples{
# add two sine and cosine terms, one with a period of 12 (for months), the other with a period of 52 (for weeks)
list("beta * I", S, I, strata = NULL, seasonality(period = c(12, 52), s_params = c(0.01, 0.05)))
}

