% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Set the parameter values for a stochastic epidemic model object.}
\usage{
set_params(stem_object, ...)
}
\arguments{
\item{stem_object}{stochastic epidemic model object for which parameter
values should be set.}

\item{...}{}
}
\value{
stem object with updated parameters
}
\description{
Set the parameter values for a stochastic epidemic model object.
}

