% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lna.R
\name{simulate_lna}
\alias{simulate_lna}
\title{Simulate a path using the linear noise approximation}
\usage{
simulate_lna(stem_object, census_times, lna_restart, init_states)
}
\arguments{
\item{stem_object}{stochastic epidemic model object}

\item{census_times}{vector of times at which compartment counts should be
recorded. Required for \code{method = "lna"}. If supplied and \code{method
= "gillespie"}, the compartment counts at census times are returned rather
than the full paths.}

\item{lna_restart}{Either a logical indicating whether the non-restarting
version of the LNA (default) should be used, or whether the LNA should be
restarted at census times. Alternately, a vector of restart times could be
supplied.}
}
\value{
matrix containing a path simulated using the LNA
}
\description{
Simulate a path using the linear noise approximation
}

