% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_lna_paths}
\alias{simulate_lna_paths}
\title{Simulate a data matrix from the measurement process of a stochastic epidemic
model.}
\usage{
simulate_lna_paths(nsim, lna_times, census_times, census_inds, restart_inds,
  lna_pars, init_states, incidence_codes, lna_pointer, set_pars_pointer)
}
\arguments{
\item{nsim}{integer for the number of simulations}

\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{census_times}{vector of times at which the LNA should be censused}

\item{census_inds}{logical vector the same length as lna_times indicating
whether the LNA path is censused at each of the lna times}

\item{restart_inds}{logical vector the same length as lna_times indicating
whether the LNA is to be restarted at each of the lna times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_states}{matrix of initial state vectors}

\item{incidence_codes}{vector of incidence codes, or 0 if no incidence
compartments}

\item{lna_pointer}{external pointer to LNA integration fcn}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
array of paths simulated using the LNA.
}
\description{
Simulate a data matrix from the measurement process of a stochastic epidemic
model.
}

