% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_r_measure}
\alias{simulate_r_measure}
\title{Simulate a data matrix from the measurement process of a stochastic epidemic
model.}
\usage{
simulate_r_measure(censusmat, measproc_indmat, parameters, constants, tcovar,
  r_measure_ptr)
}
\arguments{
\item{censusmat}{matrix of compartment counts at observation times}

\item{measproc_indmat}{logical matrix for which measure variables are
observed at which times}

\item{parameters}{numeric vector of model parameters}

\item{constants}{numeric vector of constants}

\item{tcovar}{numeric matrix of time-varying covariate values at observation
times.}

\item{r_measure_ptr}{external pointer to measurement process simulation fcn}
}
\value{
matrix with a simulated dataset from a stochastic epidemic model.
}
\description{
Simulate a data matrix from the measurement process of a stochastic epidemic
model.
}
