% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lna.R
\name{solve_lna}
\alias{solve_lna}
\title{Solve the LNA ODEs over a single time interval}
\usage{
solve_lna(stemr_lnamod, det_proc, innov_proc, t_L, t_R, n_comps)
}
\arguments{
\item{stemr_lnamod}{list stemr LNA model objects}

\item{det_proc}{vector for the deterministic process}

\item{innov_proc}{vector for the stochastic term}

\item{t_L}{left endpoint of the interval}

\item{t_R}{right endpoint of the interval}

\item{n_comps}{number of model compartments}
}
\value{
list containing the solution to the LNA over the time interval
  [t_L,t_R].
}
\description{
Solve the LNA ODEs over a single time interval
}

