% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_measproc_objects.R
\name{stan_measproc_objects}
\alias{stan_measproc_objects}
\title{Generates additional objects for use in constructing LNA user defined functions for use in Stan.}
\usage{
stan_measproc_objects(dynamics, meas_procs, obsmat)
}
\arguments{
\item{dynamics}{list of parsed dynamics}

\item{meas_procs}{list of parsed measurment process functions}

\item{obstimes}{list of vectors of observation times}
}
\value{
list of additional objects for Stan user defined functions
}
\description{
Generates additional objects for use in constructing LNA user defined functions for use in Stan.
}
