% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_dynamics.R
\name{stem_dynamics}
\alias{stem_dynamics}
\title{Generate the objects governing the dynamics of a stochastic epidemic model.}
\usage{
stem_dynamics(rates, parameters, compartments, strata = NULL,
  constants = NULL, timevar = NULL, adjacency = NULL)
}
\arguments{
\item{rates}{list of rate lists, see discussion below for instructions on
specifying each rate list.}

\item{parameters}{character vector of parameter names}

\item{compartments}{character vector of compartment names if there is a
single stratum, or if there are multiple strata a list of character vectors
where the name of each character vector is the compartment name and the
character vector lists the strata to in which the compartment exists. The
reserved word "ALL" can be used instead of listing all strata.}

\item{strata}{vector of stratum names, required only if not all compartments
  are common to all strata

  (e.g. compartments = list(S = "ALL", I = "ALL", R = "ALL", D = "old");
  strata = c("infants", "young", "old");}

\item{constants}{optional character vector of constants that are referenced
in the rate functions}

\item{timevar}{string specifying the name of the time variable, required if
the time variable is referenced in one of the rates}

\item{adjacency}{optional matrix specifying the adjacency structure of
strata, with 0 entries indicating non-adjacency and 1 for adjacency. Rows
and columns must be labeled.}
}
\value{
list with evaluated rate functions and objects for managing the
  bookkeeping for epidemic paths.
}
\description{
Generate the objects governing the dynamics of a stochastic epidemic model.
}
\section{Specifying the rate functions}{
 Each rate function should be provided
  as a list with four character strings - the rate function, the compartment
  from which an individual exits, the compartment that she enters, and,
  optionally, the strata to which the rate function applies: \code{list(rate,
  from, to, strata = NULL)}. The \code{from} and \code{to} arguments should
  be strings that match entries in the \code{compartments} argument. The
  optional \code{strata} argument can either be a character vector of strata
  for which the rate applies, or may be specified as "ALL" if the rate
  function applies to common compartments in all strata. If there are
  multiple strata, the strata argument for each rate is required.

  \emph{VERY IMPORTANT:} The rate functions must be specified at the subject
  level as they are parsed internally and converted to lumped rates. For
  example, in a standard SIR model, the subject-level infectivity and
  recovery rates are \eqn{\beta * I} and \eqn{\mu}, whereas the lumped rates
  which will be generated automatically are \eqn{\beta * I * S} and \eqn{\mu
  * I}.

  The rate functions should be provided as string snippets of valid C++ code.
  Each function is parsed internally for references to parameters,
  compartments, strata, constants, and the time variable. There are a few
  case-sensitive reserved words that are provided to facilitate the
  specification of rate functions when there are multiple strata: "ALL",
  "ADJ", "SELF".  In an SIR model with multiple strata, for example, we could
  specify infectivity rates as follows: \enumerate{ \item list("beta *
  I_SELF", "S", "I", "ALL"): each susceptible contacts the infecteds within
  her own stratum at rate \eqn{\beta * I}, but does not contract an infection
  from outside her stratum. This rate applies to subjects in all strata.
  \item list("beta * pow(I_SELF, alpha)", "S", "I", "ALL"): each susceptible
  contacts the infecteds within her own stratum at rate \eqn{\beta *
  I^\alpha}, but does not contract an infection from outside her stratum.
  Note that the rate must be specified as valid C++, so exponentiation is
  done using the \code{pow()} function. \item list("beta * sum(I_ALL)", "S",
  "I", "ALL"): each susceptible contacts all infecteds in the population at
  rate \eqn{\beta * \sum_s I_s}, regardless of stratum. Note that 'I_ALL'
  will be replaced by a vector of I_strata, therefore using 'I_all' outside
  of a function, e.g. beta * I_ALL, will result in an error. However,
  beta*sum(I_ALL) is well defined. \item list("beta1 * I_SELF + beta2 *
  sum(I_adj)", "S", "I", "ALL"): each susceptible contacts infecteds in her
  own stratum at rate \eqn{\beta1 * I_SELF}, and contacts infecteds in
  adjacent strata (specified in the adjacency matrix) at rate \eqn{\beta *
  \sum_{s: stratum 's' is adjacent} I_s}. Note that in this case, the
  diagonal entries in the adjacency matrix should be set to zero.}

  The \code{from} and \code{to} arguments will automatically be updated to
  reflect the strata if there are multiple strata and if no strata are
  specified in those arguments. Suppose, for example, that we define strata
  to be two adjacent regions, east and west. The first rate given above will
  be understood as governing transitions from S_east to I_east, and from
  S_west to I_west. If we we want to define rates for transitions where a
  subject crosses strata, we can do so by including the strata in the
  \code{from} and \code{to} arguments. \enumerate{\item list("beta * I",
  "S_east","I_west", "east"): each susceptible in the east contacts infected
  individuals in the east at rate \eqn{\beta * I_east}, and then transitions
  to the infected compartment in the west stratum. \item list("beta *
  I_SELF", "S", "I_west", "ALL"): all susceptibles contact infecteds in their
  own region at rate \eqn{\beta * I}, but upon becoming infected end up in
  the infected compartment in the west stratum.}
}

