% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference.R
\name{stem_inference}
\alias{stem_inference}
\title{Perform inference for a stochastic epidemic model using either the linear
noise approximation or agent-based Bayesian data augmentation.}
\usage{
stem_inference(stem_object, method, iterations, priors, kernels,
  thin_params = 1, thin_latent_proc = ceiling(iterations/100),
  messages = TRUE)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "lna" or "bda"}

\item{iterations}{number of iterations}

\item{priors}{list of priors for model parameters, each generated by a call
to the \link{\code{prior}} function. Should not include priors for the
initial compartment counts, which are specified in the
\code{\link{stem_initializer}} function.}

\item{kernels}{list of MCMC transition kernels, each generated by a call to
the \link{\code{kernel}} function}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100) so that every 100th path will be saved}

\item{messages}{should status messages be printed? defaults to TRUE.}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Perform inference for a stochastic epidemic model using either the linear
noise approximation or agent-based Bayesian data augmentation.
}

