% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference_lna.R
\name{stem_inference_lna}
\alias{stem_inference_lna}
\title{Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.}
\usage{
stem_inference_lna(stem_object, iterations, priors, MCMC_kernel, t0_kernel,
  thin_params, thin_latent_proc, initialization_attempts = 500,
  n_ess_updates = 1, messages)
}
\arguments{
\item{stem_object}{stochastic epidemic model object with model dynamics, the
measurement process, and a dataset.}

\item{iterations}{number of MCMC iterations}

\item{priors}{either a list of priors, each of which is constructed using the
\code{prior} function, or a list of functions for computing the prior
density as well as transforming parameters to and from their estimation
scales}

\item{MCMC_kernel}{list containing the MCMC_kernel method, proposal covariance
matrix, and an external pointer for the compiled MCMC_kernel function}

\item{t0_kernel}{output of \link{\code{t0_kernel}}, specifying the RWMH
transition MCMC_kernel for t0 and the truncated normal distribution prior.}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100)}

\item{initialization_attempts}{number of attempts to initialize the latent
path before breaking.}

\item{n_ess_updates}{number of elliptical slice sampling updates per iteration}

\item{messages}{should status messages be generated in an external text file?
If so, the iteration number is printed every 1000th iteration. defaults to
TRUE.}
}
\value{
list with parameter posterior samples and MCMC diagnostics
}
\description{
Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.
}
