% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_measure.R
\name{stem_measure}
\alias{stem_measure}
\title{Generate a list of objects governing the measurement process for a stochastic
epidemic model.}
\usage{
stem_measure(emissions, dynamics, data = NULL, messages = TRUE)
}
\arguments{
\item{emissions}{list of emmision lists, each generated by a call to the
\code{\link{emission}} function.}

\item{dynamics}{processed list of objects governing the model dynamics,
returned by the \code{\link{stem_dynamics}} function.}

\item{data}{matrix/data frame, or a list of matrices/data frames. All columns
must be named according to which compartment_strata are measured. The first
column must consist of observation times, t_1,...,t_L. If data on all
measured compartments are accrued at the same observation times, a single
matrix or data frame may be provided. If each compartment in the data was
measured at different observation times, a list of matrices or data frames
must be provided. Again, the first column of each matrix must consist of
observation times, while subsequent columns must be labeled according to
which compartment being measured.}

\item{messages}{should compilation messages be printed? defaults to true.}
}
\value{
list with evaluated measurement process functions and objects.
}
\description{
Generate a list of objects governing the measurement process for a stochastic
epidemic model.
}

