% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_ode.R
\name{stem_ode}
\alias{stem_ode}
\title{Obtain the deterministic ODE path for a stochastic epidemic model.}
\usage{
stem_ode(stem_object, census_times, init_states)
}
\arguments{
\item{stem_object}{stochastic epidemic model object}

\item{census_times}{list of times at which to evaluate the path}

\item{init_states}{initial state}
}
\value{
matrix with the ODE path
}
\description{
Obtain the deterministic ODE path for a stochastic epidemic model.
}

