% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_ode_path.R
\name{stem_ode_path}
\alias{stem_ode_path}
\title{Solve the ODEs over a single time interval}
\usage{
stem_ode_path(stemr_odemod, ode_state, t_L, t_R)
}
\arguments{
\item{stemr_odemod}{list stemr ODE model objects}

\item{ode_state}{vector for the deterministic process}

\item{t_L}{left endpoint of the interval}

\item{t_R}{right endpoint of the interval}
}
\value{
list containing the solution to the ODE over the time interval
  [t_L,t_R].
}
\description{
Solve the ODEs over a single time interval
}

