% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_parameters.R
\name{stem_parameters}
\alias{stem_parameters}
\title{Assign a vector of parameters to a stem object}
\usage{
stem_parameters(parameters, stem_object)
}
\arguments{
\item{parameters}{named vector of parameters}

\item{stem_object}{stem list}
}
\value{
stem_object with new parameters
}
\description{
Assign a vector of parameters to a stem object
}
