% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_comp_fcns.R
\name{sub_comp_fcns}
\alias{sub_comp_fcns}
\title{Makes comp_fcn replacements.}
\usage{
sub_comp_fcns(rate_string, comp, subs)
}
\arguments{
\item{rate_string}{rate string in which to make replacements}

\item{comp}{character string to be replaced}

\item{subs}{character vector of compartments to be substituted}
}
\value{
rate string
}
\description{
Makes comp_fcn replacements.
}

