% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_powers.R
\name{sub_powers}
\alias{sub_powers}
\title{Parse a string and substitute powers of the form a^b with pow(a,b).}
\usage{
sub_powers(string)
}
\arguments{
\item{string}{}
}
\value{
string suitable for C++
}
\description{
The function for accomplishing this was proposed by Roland on the following
\href{https://stackoverflow.com/questions/40606723/substitute-the-power-symbol-with-cs-pow-syntax-in-mathematical-expression/40612497}{stack
overflow thread.}
}
