% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_kernel.R
\name{t0_kernel}
\alias{t0_kernel}
\title{Specify the prior for t0 and the standard deviation of the random--walk
proposal distribution.}
\usage{
t0_kernel(rw_sd, mean, sd, lower = NULL, upper = NULL)
}
\arguments{
\item{rw_sd}{standard deviation of the random walk proposal.}

\item{mean}{prior mean of the truncated normal distribution}

\item{sd}{prior standard deviation of the truncated normal distribution}

\item{lower}{lower bound, defaults to -Inf if not specified.}

\item{upper}{upper bound, defaults to the first observation time if not
specified.}
}
\value{
arguments for specifying prior and proposal distribution for t0
}
\description{
Specify the prior for t0 and the standard deviation of the random--walk
proposal distribution.
}
