% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_estimation_scale}
\alias{to_estimation_scale}
\title{Transform the parameters from their natural scale to the estimation scale}
\usage{
to_estimation_scale(natural_params, scaled_params, scales)
}
\arguments{
\item{natural_params}{numeric vector of parameter values on their natural
scale}

\item{scaled_params}{numeric vector of scaled parameter values, to be
modified in pleace}

\item{scales}{character vector of estimation scales, either 'linear', 'log',
or 'logit'}
}
\value{
updates the scaled_params vector in place with the scaled parameter
  values
}
\description{
Transform the parameters from their natural scale to the estimation scale
}

