% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpar_control.R
\name{tpar_control}
\alias{tpar_control}
\title{Generates a list of settings for sampling the latent LNA paths and
time-varying parameters via elliptical slice sampling.}
\usage{
tpar_control(
  n_updates = 1,
  bracket_width = 2 * pi,
  bracket_update_iter = Inf,
  bracket_scaling = 2 * sqrt(2 * log(10))
)
}
\arguments{
\item{n_updates}{number of elliptical slice sampling updates per MCMC
iteration, defaults to 1.}

\item{bracket_width}{width of the elliptical slice sampling brackets, must
be in (0,2*pi], and default to 2*pi.}

\item{bracket_update_iter}{iteration at which the widths of the elliptical
slice sampling brackets should be shrunk, defaults to Inf and the bracket
widths are kept constant.}

\item{bracket_scaling}{Scaling factors for elliptical slice sampling bracket
widths. If brackets are to be shrunk, the new width is set to the minimum
of 2*pi or the scaling multiplied by the standard deviation of the previous
ESS angles. The scalings default to the full width at one tenth maximum for
a gaussian.}
}
\value{
list with settings for elliptical slice sampling
}
\description{
Generates a list of settings for sampling the latent LNA paths and
time-varying parameters via elliptical slice sampling.
}
