% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_factors}
\alias{update_factors}
\title{Update slice factor directions for automated factor slice sampling}
\usage{
update_factors(slice_singvals, slice_factors, slice_factors_t, kernel_cov)
}
\arguments{
\item{slice_singvals}{vector of singular values}

\item{slice_factors}{vector of singular vectors}

\item{slice_factors_t}{transpost matrix for singular vectors}

\item{kernel_cov}{empirical covariance matrix of model params}
}
\value{
update eigenvalues and eigenvectors in place
}
\description{
Update slice factor directions for automated factor slice sampling
}
