% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_initdist_ode.R
\name{update_initdist_ode}
\alias{update_initdist_ode}
\title{Update initial compartment volumes via elliptical slice sampling.}
\usage{
update_initdist_ode(initdist_objects, init_volumes_cur, init_volumes_prop,
  path_cur, data, ode_parameters, ode_param_vec, tparam, pathmat_prop,
  censusmat, emitmat, flow_matrix, stoich_matrix, ode_times, forcing_inds,
  forcing_matrix, ode_param_inds, ode_const_inds, ode_tcovar_inds,
  ode_initdist_inds, param_update_inds, census_indices, ode_event_inds,
  measproc_indmat, ode_pointer, ode_set_pars_pointer, d_meas_pointer,
  do_prevalence, step_size, initdist_steps, initdist_angle,
  initdist_bracket_width, n_initdist_updates)
}
\arguments{
\item{path_cur}{list with the current path}

\item{data}{matrix containing the dataset}

\item{tparam}{list containing the time-varying parameters}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{LNA stoichiometry matrix}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{n_ess_updates}{number of elliptical slice sampling updates}
}
\value{
updated time-varying parameter values and lna path
}
\description{
Update initial compartment volumes via elliptical slice sampling.
}
