% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_interval_widths}
\alias{update_interval_widths}
\title{Update factors and interval widths for automated factor slice sampling}
\usage{
update_interval_widths(
  interval_widths,
  n_expansions_afss,
  n_contractions_afss,
  c_expansions_afss,
  c_contractions_afss,
  slice_ratios,
  adaptation_factor,
  target_ratio
)
}
\arguments{
\item{interval_widths}{vector of interval widths}

\item{n_expansions_afss}{vector with number of expansion}

\item{n_contractions_afss}{vector with number of contractions}

\item{c_expansions_afss}{cumulative numbers of expansions}

\item{c_contractions_afss}{cumulative numbers of contractions}

\item{slice_ratios}{vector for storing ratio of cumulative number of 
expansions over number of interval width changes}

\item{adaptation_factor}{}
}
\value{
adapt interval widths in place
}
\description{
Update factors and interval widths for automated factor slice sampling
}
