% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lna_path.R
\name{update_lna_path}
\alias{update_lna_path}
\title{Sample a new LNA path via elliptical slice sampling.}
\usage{
update_lna_path(path_cur, lna_parameters, stoich_matrix, lna_ess_pointer,
  lna_times, initdist_parameters, fixed_inits, param_update_inds, drift_inds,
  resid_inds, log_scale, incidence_codes, data, measproc_indmat, obstime_inds,
  d_meas_pointer, parameters, constants, tcovar_censmat, do_census,
  do_incidence)
}
\arguments{
\item{path_cur}{list with the current LNA path along with its moments}

\item{stoich_matrix}{the stoichiometry matrix, i.e. the transpose of the flow
matrix.}

\item{initdist_parameters}{vector of parameters for the initial state}

\item{fixed_inits}{is the initial state fixed or random?}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{drift_inds}{vector of indices corresponding to the drift process within
the vector holding the ODEs}

\item{resid_inds}{vector of indices corresponding to the residual process
within the vector holding the ODEs}

\item{log_scale}{logical}

\item{incidence_codes}{vector of incidence codes}

\item{data}{matrix containing the data}

\item{measproc_indmat}{matrix indicating which compartments are measured}

\item{obstime_inds}{list of indices indicating at which rows each measured
variable is recorded}

\item{d_meas_pointer}{external pointer to density function}

\item{parameters}{vector of model parameters}

\item{constants}{vector of constants}

\item{tcovar_censmat}{time-varying covariates at observation times}

\item{do_census}{Does the path need to be censused?}

\item{do_incidence}{Does incidence need to be computed?}
}
\value{
list with an updated LNA path along with its ODEs
}
\description{
Sample a new LNA path via elliptical slice sampling.
}

