% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_princomps}
\alias{update_princomps}
\title{Update principal component Metropolis eigenvectors and eigenvalues}
\usage{
update_princomps(eigenvalues, eigenvectors, vec_prod_lower, comp_weights,
  kernel_cov)
}
\arguments{
\item{eigenvalues}{vector of eigenvalues of the proposal covariance matrix}

\item{eigenvectors}{matrix of eigenvectors for the proposal covariance}

\item{vec_prod_lower}{matrix for caching the inverse of the lower triangular
matrix of the eigenvector inner product}

\item{comp_weights}{vector of component weights}

\item{kernel_cov}{proposal covariance matrix}
}
\value{
update eigenvalues and eigenvectors in place
}
\description{
Update principal component Metropolis eigenvectors and eigenvalues
}
