% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tparam_ode.R
\name{update_tparam_ode}
\alias{update_tparam_ode}
\title{Update time-varying parameters via elliptical slice sampling.}
\usage{
update_tparam_ode(tparam, path_cur, data, ode_parameters, ode_param_vec,
  pathmat_prop, censusmat, emitmat, flow_matrix, stoich_matrix, ode_times,
  forcing_inds, forcing_matrix, ode_param_inds, ode_const_inds, ode_tcovar_inds,
  ode_initdist_inds, param_update_inds, census_indices, ode_event_inds,
  measproc_indmat, svd_sqrt, svd_d, svd_U, svd_V, ode_pointer,
  ode_set_pars_pointer, d_meas_pointer, do_prevalence, step_size, tparam_ess)
}
\arguments{
\item{tparam}{list containing the time-varying parameters}

\item{path_cur}{list with the current ode}

\item{data}{matrix containing the dataset}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{LNA stoichiometry matrix}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{svd_sqrt, svd_d, svd_U, svd_V}{objects for computing the SVD of LNA
diffusion matrics}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{n_ess_updates}{number of elliptical slice sampling updates}
}
\value{
updated time-varying parameter values and lna path
}
\description{
Update time-varying parameters via elliptical slice sampling.
}
