% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec2procmats}
\alias{vec2procmats}
\title{Copy the LNA ODE state vector into the drift, residual, and diffusion
processes matrices, imposing diagonal dominance for the diffusion matrix}
\usage{
vec2procmats(statevec, driftmat, residmat, diffarr, ind)
}
\arguments{
\item{statevec}{vector to be copied into, i.e. the vectorised ODE state}

\item{driftmat}{armadillo matrix for the drift process}

\item{residmat}{armadillo matrix for the residual process}

\item{diffarr}{armadillo cube for the diffusion process}

\item{ind}{row index in the ODE matrices (slice in diffusion cube)}
}
\value{
modifies the ode objects in place
}
\description{
Copy the LNA ODE state vector into the drift, residual, and diffusion
processes matrices, imposing diagonal dominance for the diffusion matrix
}

