% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec2procs}
\alias{vec2procs}
\title{Copy the LNA ODE state into the drift, residual, and diffusion processes
vectors}
\usage{
vec2procs(statevec, driftvec, residvec, diffmat)
}
\arguments{
\item{statevec}{vector to be copied into, i.e. the vectorised ODE state}

\item{driftvec}{armadillo vector for the drift process}

\item{residvec}{armadillo vector for the residual process}

\item{diffmat}{armadillo matrix for the diffusion process}
}
\value{
modifies the ode objects in place
}
\description{
Copy the LNA ODE state into the drift, residual, and diffusion processes
vectors
}

