% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_2_arr}
\alias{vec_2_arr}
\title{Copy a matrix into a column of a slice of an array}
\usage{
vec_2_arr(dest, orig, col_ind, slice_ind)
}
\arguments{
\item{dest}{array into which to copy}

\item{orig}{matrix to copy}

\item{col_ind}{column index (C++)}

\item{slice_ind}{slice index (C++)}
}
\value{
copy a matrix into an array.
}
\description{
Copy a matrix into a column of a slice of an array
}
