% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_2_mat}
\alias{vec_2_mat}
\title{Copy a vector into a matrix}
\usage{
vec_2_mat(dest, orig, ind)
}
\arguments{
\item{dest}{array into which to copy}

\item{orig}{matrix to copy}

\item{ind}{column index (C++)}
}
\value{
copy a matrix into an array.
}
\description{
Copy a vector into a matrix
}
