% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_absorbing.R
\name{which_absorbing}
\alias{which_absorbing}
\title{Detects which states in a model are absorbing states}
\usage{
which_absorbing(flow_matrix)
}
\arguments{
\item{flow_matrix}{}
}
\value{
logical vector indicating which model compartments are absorbing.
}
\description{
Detects which states in a model are absorbing states
}
