% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afss_settings.R
\name{afss_settings}
\alias{afss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
afss_settings(factor_update_interval = NULL, prob_update_interval = NULL,
  first_factor_update = 100, first_prob_update = 100,
  initial_slice_probs = NULL, initial_widths = NULL,
  n_afss_updates = NULL, sample_all_initially = TRUE,
  afss_slice_ratio = 0.5, target_prop_totsd = NULL, harss_prob = 0.05)
}
\arguments{
\item{factor_update_interval}{number of iterations between factor and
interval updates (defaults to an update every iteration). May also be
supplied as a function that takes the current iteration as an input and
returns the next iteration at which the factors should be updated. Must
have a default set so that the first update can be determined. e.g.,
\code{function(x = 10) 2*x} will first update the factors at iteration 10
and will subsequently update the factors at iterations 20, 40, 80, ...}

\item{prob_update_interval}{number of iterations between updates to the
factor sampling probabilities, defaults to the factor update interval. May
also be supplied as a function that takes the current iteration as an input
and returns the next iteration at which the factors should be updated. Must
have a default set so that the first update can be determined. The factor
sampling probabilities are $min(nugget,
sqrt(lambda_i)/sum(sqrt(lambda_i))$, where $lambda_i$ are the eigen values
of the empirical covariance matrix.}

\item{first_factor_update}{iteration at which the first factor update should
occur}

\item{first_prob_update}{iteration at which the first update to the slice
direction probabilities should occur.}

\item{initial_slice_probs}{vector of initial slice probabilities}

\item{initial_widths}{vector of initial slice widths, defaults to a vector of}

\item{n_afss_updates}{number of factor slice sampling directions to update}

\item{sample_all_initially}{should all factor directions be sampled until the
first slice probability update? defaults to TRUE}

\item{afss_slice_ratio}{target ratio of expansions / (expansions +
contractions) ones.}

\item{target_prop_totsd}{Between 0 and 1. If supplied, the number of factor
directions sampled after the first probability update set to be the maximum
of the requested number of directions and number of factors required to
explain the specified proportion of the total standard deviation in the
posterior. If not equal to the number of parameters, a hit and run update
is carried out to ensure ergodicity.}

\item{harss_prob}{probability of a hit and run update at each iteration}
}
\value{
list with additional settings for automated factor slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
