% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afss_settings.R
\name{afss_settings}
\alias{afss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
afss_settings(factor_update_interval = NULL, prob_update_interval = NULL,
  first_factor_update = 100, first_prob_update = 100,
  sample_all_initially = TRUE, n_afss_updates = NULL,
  initial_widths = NULL, initial_factors = NULL,
  initial_slice_probs = NULL, target_ratio = 0.5, use_cov = TRUE,
  tparam_update = TRUE)
}
\arguments{
\item{factor_update_interval}{number of iterations between factor and
interval updates (defaults to an update every iteration). May also be
supplied as a function that takes the current iteration as an input and
returns the next iteration at which the factors should be updated. Must
have a default set so that the first update can be determined. e.g.,
\code{function(x = 10) 2*x} will first update the factors at iteration 10
and will subsequently update the factors at iterations 20, 40, 80, ...}

\item{prob_update_interval}{number of iterations between updates to the
factor sampling probabilities, defaults to the factor update interval. May
also be supplied as a function that takes the current iteration as an input
and returns the next iteration at which the factors should be updated. Must
have a default set so that the first update can be determined. The factor
sampling probabilities are $min(nugget,
sqrt(lambda_i)/sum(sqrt(lambda_i))$, where $lambda_i$ are the singular
values of the empirical covariance matrix.}

\item{first_factor_update}{iteration at which the first factor update should
occur}

\item{first_prob_update}{iteration at which the first update to the slice
direction probabilities should occur.}

\item{sample_all_initially}{should all factors be sampled until the first
slice probability update? defaults to TRUE.}

\item{n_afss_updates}{number of factor slice sampling directions to update}

\item{initial_widths}{vector of initial slice widths}

\item{initial_factors}{matrix of initial factor loadings}

\item{initial_slice_probs}{initial slice direction probabilities}

\item{target_ratio}{target ratio of expansions/(expansions + contractions),
defaults to 0.5. Smaller values overrelax the interval widths.}

\item{use_cov}{should the slice directions be computed as singular vectors of
the covariance matrix (as opposed to the correlation)? defaults to TRUE.}

\item{tparam_update}{either TRUE if a joint hit-and-run update for
time-varying parameters should be carried out (default), or FALSE if
time-varying parameters should be updated separately.}
}
\value{
list with additional settings for automated factor slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
