% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tcovar_matrix.R
\name{build_tcovar_matrix}
\alias{build_tcovar_matrix}
\title{Construct a time-varying covariance matrix that includes time and
seasonality, based on a user-supplied time-varying covariate matrix. Used
internally.}
\usage{
build_tcovar_matrix(tcovar = NULL, timestep = NULL, t0, tmax, messages)
}
\arguments{
\item{tcovar}{user-supplied time-varying covariance matrix}

\item{timestep}{time discretization interval}

\item{t0, tmax}{left and right endpoints of the time-period over which the
process evolves}
}
\value{
time-varying covariate matrix that includes discretized intervals for
  smoothly time-varying functions. If neither a time-varying covariance
  matrix or a timestep is supplied, TCOVAR will just be a matrix with the
  left and right endpoints of the time-period.
}
\description{
Construct a time-varying covariance matrix that includes time and
seasonality, based on a user-supplied time-varying covariate matrix. Used
internally.
}
