% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CALL_INTEGRATE_STEM_LNA}
\alias{CALL_INTEGRATE_STEM_LNA}
\title{Compute the hazards by calling the hazard functions via external Xptr.}
\usage{
CALL_INTEGRATE_STEM_LNA(init, start, end, step_size, lna_ode_ptr)
}
\arguments{
\item{t}{time}

\item{state}{vector of compartment counts}

\item{parameters}{vector of model parameters}

\item{constants}{vector of constants}

\item{vector}{of time-varying covariate values}

\item{hazard_ptr}{external pointer to function used to compute the hazards}
}
\description{
Compute the hazards by calling the hazard functions via external Xptr.
}

