% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{census_lna}
\alias{census_lna}
\title{Construct a matrix containing the compartment counts and the incidence at a sequence of census times.}
\usage{
census_lna(path, census_path, census_inds, flow_matrix_lna, do_prevalence,
  init_state, incidence_codes_lna)
}
\arguments{
\item{path}{matrix containing the path to be censused.}

\item{census_path}{matrix to be filled out with the path.}

\item{flow_matrix_lna}{matrix containing the flow matrix for the LNA (no incidence)}

\item{do_prevalence}{should the prevalence be computed}

\item{init_state}{the initial compartment counts}

\item{incidence_codes_lna}{vector of column indices in the LNA path for which
incidence will be computed. This function merely records the cumulative
incidence at the census times from the LNA path.}
}
\value{
matrix containing the compartment counts at census times.
}
\description{
Construct a matrix containing the compartment counts and the incidence at a sequence of census times.
}
