% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{COMPUTE_HAZARD}
\alias{COMPUTE_HAZARD}
\title{Compute the hazards by calling the hazard functions via external Xptr.}
\usage{
COMPUTE_HAZARD(t, state, parameters, haz_ptr)
}
\arguments{
\item{t}{time}

\item{state}{vector of compartment counts and time-varying covariate values}

\item{parameters}{vector of model parameters and constants}

\item{hazard_ptr}{external pointer to function used to compute the hazards}
}
\description{
Compute the hazards by calling the hazard functions via external Xptr.
}

