% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{COMPUTE_JACOBIAN}
\alias{COMPUTE_JACOBIAN}
\title{Compute the Jacobian by calling the jacobian function via external Xptr.}
\usage{
COMPUTE_JACOBIAN(t, state, parameters, jacob_ptr)
}
\arguments{
\item{t}{time}

\item{state}{vector of compartment counts and time-varying covariate values}

\item{parameters}{vector of model parameters and constants}

\item{jacobian_ptr}{external pointer to function used to compute the jacobian}
}
\description{
Compute the Jacobian by calling the jacobian function via external Xptr.
}

