% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_settings.R
\name{ess_settings}
\alias{ess_settings}
\title{Generates a list of settings for sampling the latent LNA paths via elliptical
slice sampling.}
\usage{
ess_settings(n_ess_updates = 1, ess_schedule = NULL,
  randomize_schedule = TRUE)
}
\arguments{
\item{n_ess_updates}{number of elliptical slice sampling updates per MCMC
iteration, defaults to 1.}

\item{ess_schedule}{If \code{NULL}, the LNA paths for all elementary
transitions will be sampled jointly (default). May also be specified as
"bystratum" if stratum specific transitions should be proposed one at a
time, or as a list of charcter vectors for names of elementary transition
events (i.e. each character vector corresponds to the names of rows in an
LNA flow matrix) e.g., \code{"bystratum"} and \code{list(c("S_male2Imale",
"I_male2R_male"), c("S_female2I_female", "I_female2R_female"))} could both
be used to sample the LNA paths for males and females in a sex stratified
SIR model. \code{"bystratum"} is coerced internally into the list.}

\item{randomize_order}{if \code{TRUE} (default) and the \code{ess_schedule}
contains more than one vector, then the LNA paths for the events in the
vectors are sampled in a random order in each MCMC iteration (e.g., males
and females are sampled in a random order).}
}
\description{
Generates a list of settings for sampling the latent LNA paths via elliptical
slice sampling.
}
