% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_d_measure}
\alias{evaluate_d_measure}
\title{Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.}
\usage{
evaluate_d_measure(emitmat, obsmat, statemat, measproc_indmat, parameters,
  constants, tcovar_censusmat, d_meas_ptr)
}
\arguments{
\item{emitmat}{matrix of emission probabilities}

\item{obsmat}{matrix containing the data}

\item{statemat}{matrix containing the compartment counts at the observation
times}

\item{measproc_indmat}{logical matrix indicating which compartments are
observed at every observation time}

\item{parameters}{numeric vector of parameter values}

\item{constants}{numeric vector of constants}

\item{tcovar_censusmat}{numeric vector of time-varying covariate values}

\item{d_meas_ptr}{external pointer to measurement process density function}
}
\description{
Evaluate the log-density of the measurement process by calling measurement
process density functions via external Xptr.
}
