% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_d_measure_LNA}
\alias{evaluate_d_measure_LNA}
\title{Evaluate the log-density of a possibly time-verying measurement process
by calling measurement process density functions via external Xptr.}
\usage{
evaluate_d_measure_LNA(emitmat, obsmat, censusmat, measproc_indmat,
  lna_parameters, lna_params_temp, lna_param_inds, lna_const_inds,
  lna_tcovar_inds, param_update_inds, census_indices, d_meas_ptr)
}
\arguments{
\item{emitmat}{matrix of emission probabilities}

\item{obsmat}{matrix containing the data}

\item{censusmat}{matrix containing the state of the latent process at
observation times}

\item{measproc_indmat}{logical matrix indicating which compartments are
observed at every observation time}

\item{lna_parameters}{matrix containing the LNA parameters, constants and
time-varying coariates.}

\item{lna_params_temp}{container for storing the LNA parameters at each
observation time.}

\item{lna_param_inds}{indices for the model parameters.}

\item{lna_const_inds}{indices for the constants.}

\item{lna_tcovar_inds}{indices for the time-varying covariates.}

\item{param_update_inds}{logical vector indicating when model parameters
should be updated.}

\item{census_indices}{vector of indices when the LNA path has been censused.}

\item{d_meas_ptr}{external pointer to measurement process density function}
}
\description{
Evaluate the log-density of a possibly time-verying measurement process
by calling measurement process density functions via external Xptr.
}
