% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stem.R
\name{fit_stem}
\alias{fit_stem}
\title{Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.}
\usage{
fit_stem(
  stem_object,
  method,
  mcmc_kern,
  iterations,
  initialization_attempts = 500,
  ess_warmup = 50,
  thinning_interval = 100,
  return_adapt_rec = FALSE,
  return_ess_rec = FALSE,
  print_progress = 0,
  status_filename = NULL
)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "lna" or ode.}

\item{mcmc_kern}{MCMC transition kernel generated by a call to the
\code{mcmc_kernel} function.}

\item{iterations}{number of iterations}

\item{initialization_attempts}{number of initialization attempts}

\item{ess_warmup}{number of preliminary ESS iterations for the LNA, initial
conditions, and time varying parameters prior to starting MCMC}

\item{thinning_interval}{thinning interval for posterior samples, defaults to
saving every 100th sample}

\item{return_adapt_rec}{should the MCMC samples be returned during
adaptation? defaults to FALSE.}

\item{return_ess_rec}{should elliptical slice sampling steps and angles be
returned? defaults to FALSE}

\item{print_progress}{interval at which to print progress to a text file. If
0 (default) progress is not printed.}

\item{status_filename}{string to pre-append to status files, defaults to LNA
or ODE depending on the method used.}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.
}
