% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stem.R
\name{fit_stem}
\alias{fit_stem}
\title{Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.}
\usage{
fit_stem(
  stem_object,
  method,
  mcmc_kern,
  iterations,
  initialization_attempts = 500,
  thinning_interval = ceiling(iterations/100),
  return_adapt_rec = FALSE,
  print_progress = 0,
  status_filename = NULL,
  messages = FALSE
)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "lna" or ode.}

\item{mcmc_kern}{MCMC transition kernel generated by a call to the
\code{mcmc_kernel} function.}

\item{iterations}{number of iterations}

\item{initialization_attempts}{number of initialization attempts}

\item{thinning_interval}{thinning interval for posterior samples, defaults to
ceiling(iterations/100) so that every 100th sample will be saved}

\item{return_adapt_rec}{should the MCMC adaptation record be returned?
defaults to FALSE.}

\item{print_progress}{interval at which to print progress to a text file. If
0 (default) progress is not printed.}

\item{status_filename}{string to pre-append to status files, defaults to LNA
or ODE depending on the method used.}

\item{messages}{should status messages be printed? defaults to FALSE.}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Fit a stochastic epidemic model using the linear noise approximation or
ordinary differential equations to approximate the latent epidemic process.
}
