% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harss_settings.R
\name{harss_settings}
\alias{harss_settings}
\title{Generate a list of settings for automated factor slice sampling}
\usage{
harss_settings(n_harss_updates = 1, initial_width = 1, target_ratio = 0.5,
  tparam_update = TRUE)
}
\arguments{
\item{n_harss_updates}{number of hit-and-run updates.}

\item{initial_width}{initial bracket width, defaults to 1.}

\item{target_ratio}{target ratio of expansions/(expansions + contractions),
defaults to 0.5. Smaller values overrelax the interval widths.}

\item{tparam_update}{either TRUE if a joint hit-and-run update for
time-varying parameters should be carried out (default), or FALSE if
time-varying parameters should be updated separately.}
}
\value{
list with additional settings for hit and run slice sampling
}
\description{
Generate a list of settings for automated factor slice sampling
}
