% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_lna.R
\name{initialize_lna}
\alias{initialize_lna}
\title{Initialize the LNA path}
\usage{
initialize_lna(data, lna_parameters, censusmat, emitmat, stoich_matrix,
  lna_pointer, lna_set_pars_pointer, lna_times, lna_param_inds, lna_const_inds,
  lna_tcovar_inds, lna_initdist_inds, param_update_inds, census_indices,
  lna_event_inds, measproc_indmat, d_meas_pointer, do_prevalence, forcing_inds,
  forcing_matrix, initialization_attempts, step_size)
}
\arguments{
\item{data}{matrix containing the dataset}

\item{lna_parameters}{parameters, contants, time-varying covariates at LNA
times}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{LNA stoichiometry matrix}

\item{lna_pointer}{external LNA pointer}

\item{lna_set_pars_pointer}{pointer for setting the LNA parameters}

\item{lna_times}{times at whicht eh LNA should be evaluated}

\item{lna_param_inds}{C++ column indices for parameters}

\item{lna_const_inds}{C++ column indices for constants}

\item{lna_tcovar_inds}{C++ column indices for time varying covariates}

\item{lna_initdist_inds}{C++ column indices in the LNA parameter matrix for
the initial state}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{lna_event_inds}{vector of column indices in the LNA path for which
incidence will be computed.}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{initialization_attempts}{number of initialization attempts}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}
}
\value{
LNA path along with its stochastic perturbations
}
\description{
Initialize the LNA path
}
