% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_lna.R
\name{initialize_lna}
\alias{initialize_lna}
\title{Initialize the LNA path}
\usage{
initialize_lna(data, lna_parameters, censusmat, emitmat, stoich_matrix,
  lna_pointer, lna_set_pars_pointer, lna_times, lna_params_temp, lna_param_inds,
  lna_const_inds, lna_tcovar_inds, lna_initdist_inds, param_update_inds,
  incidence_codes, census_incidence_codes, census_indices, measproc_indmat,
  obstime_inds, d_meas_pointer, do_prevalence, initialization_attempts)
}
\arguments{
\item{data}{matrix containing the dataset}

\item{lna_parameters}{parameters, contants, time-varying covariates at LNA
times}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{LNA stoichiometry matrix}

\item{lna_pointer}{external LNA pointer}

\item{lna_set_pars_pointer}{pointer for setting the LNA parameters}

\item{lna_times}{times at whicht eh LNA should be evaluated}

\item{lna_initdist_inds}{C++ column indices in the LNA parameter matrix for
the initial state}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{incidence_codes}{C++ indices for LNA count process compartments for
which incidence is to be computed}

\item{census_incidence_codes}{C++ indices for columns in the census matrix
containing incidence data}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{obstime_inds}{list of C++ indices indicating when each measurement
process is evaluated}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{initialization_attempts}{number of initialization attempts}

\item{parameters}{vector of model parameters}

\item{constants}{vector of model constants}

\item{tcovar_censmat}{time-varying covariates at LNA times}
}
\value{
LNA path along with its stochastic perturbations
}
\description{
Initialize the LNA path
}
