% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Construct a list for building a transition kernel to be used in proposing new
parameter values.}
\usage{
kernel(parameters, method = "mvn_rw", cov_mtx)
}
\arguments{
\item{parameters}{character vector with names of parameters}

\item{method}{algorithm to be used in generating parameter updates -
currently only multivariate normal random walk is supported.}

\item{cov_mtx}{variance-covariance matrix for random walk proposals, with
rows and columns corresponding to the parameters argument}
}
\value{
list containing objects to be used in the proposals.
}
\description{
Construct a list for building a transition kernel to be used in proposing new
parameter values.
}

