% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_draws_2_lna}
\alias{map_draws_2_lna}
\title{Map N(0,1) stochastic perturbations to an LNA path.}
\usage{
map_draws_2_lna(pathmat, draws, lna_times, lna_pars, init_start,
  param_update_inds, stoich_matrix, svd_sqrt, svd_d, svd_U, svd_V, lna_pointer,
  set_pars_pointer)
}
\arguments{
\item{pathmat}{matrix where the LNA path should be stored}

\item{draws}{matrix of N(0,1) draws to be mapped to an LNA path}

\item{lna_times}{vector of interval endpoint times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{svd_sqrt}{matrix in which to compute the matrix square root of the LNA
covariance matrices}

\item{svd_d}{vector in which to store SVD singular values}

\item{svd_U}{matrix in which to store the U matrix of the SVD}

\item{svd_V}{matrix in which to store the V matrix of the SVD}

\item{lna_pointer}{external pointer to LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
fill out pathmat with the LNA path corresponding to the stochastic
  perturbations.
}
\description{
Map N(0,1) stochastic perturbations to an LNA path.
}
