% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_matching.R
\name{maximize_loglik_ode}
\alias{maximize_loglik_ode}
\title{Obtain estimates of stochastic epidemic model (SEM) parameters that maximize
the observed data likelihood for a SEM approximated by its deterministic ODE
limit.}
\usage{
maximize_loglik_ode(stem_object, bounds)
}
\arguments{
\item{stem_object}{stem_object with}

\item{bounds}{list containing named vectors of lower and upper bounds for the
model parameters and initial compartment volumes, e.g., \code{list(upper =
c(beta = 1, mu = 1, S_0 = 99, I_0 = 10, R_0 = 10), lower = c(beta = 0, mu =
0, S_0 = 90, I_0 = 1, R_0 = 0))}.}

\item{method}{optimization method to be passed to \code{optim}.}
}
\value{
list containing the parameter estimates and the inverse of the
  Hessian.
}
\description{
NOTE: This should not be considered a reliable method for inference for SEMs.
Rather, it is perhaps useful as a method for obtaining initial values for
other methods.
}
