% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_g_adaptive}
\alias{mvn_g_adaptive}
\title{Global Metropolis random walk with global adaptive scaling}
\usage{
mvn_g_adaptive(params_prop, params_cur, kernel_cov, proposal_scaling, nugget)
}
\arguments{
\item{params_prop}{vector in which the proposed parameters should be stored}

\item{params_cur}{vector containing the current parameter vector}

\item{kernel_cov}{vector of component proposal standard deviations}

\item{proposal_scaling}{scaling parameter for the proposal}

\item{nugget}{fixed covariance nugget contribution}
}
\value{
propose new parameter values in place
}
\description{
Global Metropolis random walk with global adaptive scaling
}
