% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parblock.R
\name{parblock}
\alias{parblock}
\title{Define a parameter block for an MCMC kernel}
\usage{
parblock(
  pars_nat,
  pars_est,
  priors,
  alg,
  sigma,
  initializer = NULL,
  control = NULL
)
}
\arguments{
\item{pars_nat}{character vector of parameter names on their natural scales
(i.e., corresponding to parameters supplied to stem_dynamics)}

\item{pars_est}{character vector of parameter names on their estimation scale
(i.e., corresponding to row/column names in the kernel covariance matrix).
IMPORTANT: the estimation scale for a parameter should be unconstrained,
e.g., if mu is the recovery rate in an SIR model, it is natural to use
the log infectious period duration, -log(mu) as the estimation scale.}

\item{priors}{A list of three functions supplied by the user with names
"logprior", "to_estimation_scale", and "from_estimation_scale" (N.B. All
three must be supplied). The first of these functions should take as an
argument a numeric vector of parameters on their estimation scales. The
functions for converting parameters to and from their estimation scales.
The priors should not include priors for the initial compartment counts or
time-varying parameters.}

\item{alg}{either "mvnmh"  or "mvnss" for multivariate normal metropolis
hastings updates or multivariate normal slice sampling updates,
respectively.}

\item{sigma}{initial covariance matrix for the parameter block, possibly to
be adapted.}

\item{initializer}{optional function for initializing the parameters in the
parameter block}

\item{control}{list of mcmc control settings, generated by a call to
\code{mvnmh_control} or \code{mvnss_control} as appropriate.}
}
\value{
parameter block for use in MCMC kernel
}
\description{
Define a parameter block for an MCMC kernel
}
