% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parblock.R
\name{parblock}
\alias{parblock}
\title{Define a parameter block for an MCMC kernel}
\usage{
parblock(pars)
}
\arguments{
\item{pars}{either a numeric vector of parameter indices or a character
vector of parameter names on their estimation scale (i.e., corresponding to
row/column names in the kernel covariance matrix)}
}
\value{
parameter block for use in MCMC kernel
}
\description{
Define a parameter block for an MCMC kernel
}
