% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rates.R
\name{parse_rates}
\alias{parse_rates}
\title{Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.}
\usage{
parse_rates(rates, param_codes, compartment_codes, const_codes, tcovar_codes,
  allrates = FALSE)
}
\arguments{
\item{rates}{list of rate functions}

\item{param_codes}{vector of parameter codes}

\item{compartment_codes}{vector of compartment codes}

\item{const_codes}{vector of constant names}

\item{tcovar_codes}{vector of time-varying covariate codes}

\item{allrates}{compile each individual rate function, defaults to FALSE}
}
\value{
Two vector of strings that serve as function pointers.
}
\description{
Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.
}

