% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rates.R
\name{parse_rates}
\alias{parse_rates}
\title{Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.}
\usage{
parse_rates(rates, messages = TRUE)
}
\arguments{
\item{rates}{list of rate functions}

\item{messages}{logical; print a message that the rates are being compiled}
}
\value{
Two vector of strings that serve as function pointers.
}
\description{
Instatiate the C++ rate functions for a stochastic epidemic model and return
a vector of function pointers.
}

