% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna}
\alias{propose_lna}
\title{Simulate a data matrix from the measurement process of a stochastic epidemic
model.}
\usage{
propose_lna(lna_times, lna_pars, param_update_inds, flow_matrix, lna_pointer,
  set_pars_pointer, enforce_monotonicity)
}
\arguments{
\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{flow_matrix}{original flow matrix giving the changes to compartments
from each reaction}

\item{lna_pointer}{external pointer to LNA integration fcn}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}

\item{enforce_monotonicity}{should monotinicity in the sample paths of the
LNA counting processes be enforced?}
}
\value{
array with slices containing the LNA path, the path of the residual
  process, the drift process, the residual process (conditional means), and
  diffusion process.
}
\description{
Simulate a data matrix from the measurement process of a stochastic epidemic
model.
}

