% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna}
\alias{propose_lna}
\title{Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.}
\usage{
propose_lna(lna_times, lna_pars, init_start, param_update_inds, stoich_matrix,
  forcing_inds, forcing_matrix, reject_negatives, max_attempts, step_size,
  lna_pointer, set_pars_pointer)
}
\arguments{
\item{lna_times}{vector of interval endpoint times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{forcing_matrix}{matrix containing the forcings.}

\item{reject_negatives}{logical for whether negative increments or compartment 
volumes should lead to rejection of the sampled path. If false, draws that lead 
to either are re-sampled in place instead of throwing an error. N.B. Resampling 
targets the WRONG distribution and should only be used when initializing an LNA 
path with the understanding that further warm-up under the correct distribution 
is required.}

\item{max_attempts}{maximum number of tries to restart negative increments if 
reject_negatives is false.}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{lna_pointer}{external pointer to the compiled LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting LNA pars.}
}
\value{
list containing the stochastic perturbations (i.i.d. N(0,1) draws) and
the LNA path on its natural scale which is determined by the perturbations.
}
\description{
Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.
}
