% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna}
\alias{propose_lna}
\title{Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.}
\usage{
propose_lna(
  lna_times,
  lna_draws,
  lna_pars,
  lna_param_inds,
  lna_tcovar_inds,
  init_start,
  param_update_inds,
  stoich_matrix,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  max_attempts,
  step_size,
  lna_pointer,
  set_pars_pointer
)
}
\arguments{
\item{lna_times}{vector of interval endpoint times}

\item{lna_draws}{vector of N(0,1) draws to be mapped to the path}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{max_attempts}{maximum number of tries if the first increment is rejected}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{lna_pointer}{external pointer to the compiled LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting LNA pars.}

\item{forcing_matrix}{matrix containing the forcings.}
}
\value{
list containing the stochastic perturbations (i.i.d. N(0,1) draws) and
the LNA path on its natural scale which is determined by the perturbations.
}
\description{
Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.
}
