% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propose_lna}
\alias{propose_lna}
\title{Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.}
\usage{
propose_lna(lna_times, lna_pars, init_start, param_update_inds, stoich_matrix,
  lna_pointer, set_pars_pointer)
}
\arguments{
\item{lna_times}{vector of interval endpoint times}

\item{lna_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the lna_times}

\item{init_start}{index in the parameter vector where the initial compartment
volumes start}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{stoich_matrix}{stoichiometry matrix giving the changes to compartments
from each reaction}

\item{lna_pointer}{external pointer to LNA integration function.}

\item{set_pars_pointer}{external pointer to the function for setting the LNA
parameters.}
}
\value{
list containing the stochastic perturbations (i.i.d. N(0,1) draws) and
the LNA path on its natural scale which is determined by the perturbations.
}
\description{
Simulate an LNA path using a non-centered parameterization for the
log-transformed counting process LNA.
}
