% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_nugget.R
\name{reset_nugget}
\alias{reset_nugget}
\title{Reset the nugget based based on a sample}
\usage{
reset_nugget(nugget, parameter_samples, start, end, estimation_scales)
}
\arguments{
\item{nugget}{current value, a scalar}

\item{parameter_samples}{parameter samples on their natural scale}

\item{start}{beginning of range to include}

\item{end}{end of range to include}

\item{estimation_scales}{character vector of estimation scales}
}
\value{
vector of nugget standard deviations
}
\description{
Reset the nugget based based on a sample
}
