% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reset_vec}
\alias{reset_vec}
\title{Reset a vector by filling it with an element}
\usage{
reset_vec(v, value = 0)
}
\arguments{
\item{v}{vector to fill with zeros}

\item{value}{to insert}
}
\value{
reset vector in place
}
\description{
Reset a vector by filling it with an element
}
