% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality}
\alias{seasonality}
\title{Produces a string for seasonal terms to be added to a rate function, supplied
as the last argument in a rate list.}
\usage{
seasonality(period = 52, common_seasonality = TRUE)
}
\arguments{
\item{period}{period of the season, defaults to 52 for weeks. May also be
supplied as a vector if S>1 in order to specify multiple terms with
different periods.}

\item{common_seasonlity}{logical for whether the seasonal terms should be
shared among strata (TRUE) or whether each stratum should have its own
seasonality}
}
\value{
string formula for the seasonal terms to be added to a rate.
}
\description{
Supplies a string for the seasonal terms \deqn{\sum \gamma_p
  \times sin(2\pi p\times t) + \delta_p \times cos(2\pi p\times t)} to be
  added to a rate function. The parameters \eqn{\gamma period} and
  \eqn{\delta period} will automatically be named to reflect whether the
  seasonal terms are common to all strata, and then appended to the parameter
  vector.
}
\examples{
# add two sine and cosine terms, one with a period of 12 (for months), the other with a period of 52 (for weeks)
list("beta * I", S, I, strata = NULL, seasonality(period = 12))
}

