% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stem.R
\name{simulate_stem}
\alias{simulate_stem}
\title{Simulations from a stochastic epidemic model.}
\usage{
simulate_stem(
  stem_object,
  nsim = 1,
  simulation_parameters = NULL,
  lna_draws = NULL,
  tparam_draws = NULL,
  tparam_values = NULL,
  paths = FALSE,
  full_paths = FALSE,
  observations = FALSE,
  method = "gillespie",
  tmax = NULL,
  census_times = NULL,
  max_attempts = 500,
  lna_method = "exact",
  lna_bracket_width = 2 * pi,
  ess_warmup = 100,
  messages = TRUE
)
}
\arguments{
\item{stem_object}{stem object list}

\item{nsim}{number of realizations to simulate}

\item{simulation_parameters}{optional list of vectors of simulation
parameters. If NULL, paths are simulated using the parameters specified in
the stem object.}

\item{lna_draws}{optional list of matrices of lna draws}

\item{tparam_draws}{optional list of lists of time-varying parameter draws}

\item{paths}{Should population-level paths at census times be returned?}

\item{full_paths}{Should complete population level paths be returned (only
for Gillespie), defaults to FALSE}

\item{observations}{Should simulated observations be returned? Requires that
a measurement process be defined in the stem object.}

\item{method}{either "gillespie" if simulating via Gillespie's direct method,
"lna" if simulating paths via the linear noise approximation, or "ode" if
simulating paths of the deterministic limit of the underlying Markov jump
process.}

\item{tmax}{the time at which simulation of the system is terminated. If not
supplied, defaults to the last observation time if not supplied.}

\item{census_times}{vector of times at which compartment counts should be
recorded.}

\item{max_attempts}{maximum number of times to attempt simulating an LNA path
before aborting due to the moments of the path being degenerate at some
point}

\item{lna_method}{defaults to "exact". If "approx", an initial path is
proposed by resampling perturbations that lead to negative increments or
volumes. The initial path is then updated via elliptical slice sampling}

\item{lna_bracket_width}{initial elliptical slice sampling bracket width to
be used if lna_method == "approx"}

\item{ess_warmup}{number of elliptical slice sampling updates before the lna
sample is saved}

\item{messages}{should a message be printed when parsing the rates?}
}
\value{
Returns a list with the simulated paths, subject-level paths, and/or
  datasets. If \code{paths = FALSE} and \code{observations = FALSE}, or if
  \code{paths = TRUE} and \code{observations = TRUE}, a list \code{nsim stem}
  paths and datasets, each returned either as a list or array, is returned.

  If \code{paths = TRUE} and \code{observations = FALSE}, a list or array of
  simulated population-level paths is returned. 

  If \code{paths = FALSE} and \code{observations = TRUE}, a list or array of
  simulated datasets is returned.
}
\description{
Simulations from a stochastic epidemic model.
}
