% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stem.R
\name{simulate_stem}
\alias{simulate_stem}
\title{Simulations from a stochastic epidemic model.}
\usage{
simulate_stem(stem_object, nsim = 1, simulation_parameters = NULL,
  paths = FALSE, full_paths = FALSE, observations = FALSE,
  subject_paths = FALSE, method = "gillespie", tmax = NULL,
  census_times = NULL, max_attempts = 500, lna_method = "exact",
  ess_warmup = 100, messages = TRUE)
}
\arguments{
\item{nsim}{number of realizations to simulate}

\item{simulation_parameters}{optional list of vectors of simulation
parameters. If NULL, paths are simulated using the parameters specified in
the stem object.}

\item{paths}{Should population-level paths at census times be returned?}

\item{full_paths}{Should complete population level paths be returned (only
for Gillespie), defaults to FALSE}

\item{observations}{Should simulated observations be returned? Requires that
a measurement process be defined in the stem object.}

\item{subject_paths}{Should population-level paths be mapped to subject-level
paths and returned (e.g. for the purpose of initializing a subject-level
collection of disease histories)? Only available for exact simulation via
the Gillespie direct method.}

\item{method}{either "gillespie" if simulating via Gillespie's direct method,
"lna" if simulating paths via the linear noise approximation, or "ode" if
simulating paths of the deterministic limit of the underlying Markov jump
process.}

\item{tmax}{the time at which simulation of the system is terminated. If not
supplied, defaults to the last observation time if not supplied.}

\item{census_times}{vector of times at which compartment counts should be
recorded.}

\item{max_attempts}{maximum number of times to attempt simulating an LNA path
before aborting due to the moments of the path being degenerate at some
point}

\item{lna_method}{defaults to "exact". If "approx", an initial path is
proposed by resampling perturbations that lead to negative increments or
volumes. The initial path is then updated via elliptical slice sampling}

\item{ess_warmup}{number of elliptical slice sampling updates before the
lna sample is saved}

\item{messages}{should a message be printed when parsing the rates?}

\item{stem}{Stochastic epidemic model object. The dynamics slot must be
supplied at a minimum. If simulations from the measurement process are
desired, the measurement process slot must also be supplied.}
}
\value{
Returns a list with the simulated paths, subject-level paths, and/or
  datasets. If \code{paths = FALSE} and \code{observations = FALSE}, or if
  \code{paths = TRUE} and \code{observations = TRUE}, a list \code{nsim stem}
  paths and datasets, each returned either as a list or array, is returned.
  If \code{subject_paths = TRUE} and \code{method = "gillespie"}, a list of
  subject-level paths is also returned.

  If \code{paths = TRUE} and \code{observations = FALSE}, a list or array of
  simulated population-level paths is returned. If \code{method =
  "gillespie"} and \code{subject_paths = TRUE}, an additional list is
  returned, containing lists of subject-level mappings. Each sublist contains
  a vector of subject states at t0, along with a mapped path.

  If \code{paths = FALSE} and \code{observations = TRUE}, a list or array of
  simulated datasets is returned.
}
\description{
Simulations from a stochastic epidemic model.
}
