% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stem.R
\name{simulate_stem}
\alias{simulate_stem}
\title{Simulations from a stochastic epidemic model.}
\usage{
simulate_stem(stem_object, nsim = 1, paths = FALSE, observations = FALSE,
  subject_paths = FALSE, method = "gillespie", t0 = NULL, tmax = NULL,
  timestep = NULL, census_times = NULL, paths_as_array = FALSE,
  datasets_as_array = FALSE, messages = TRUE)
}
\arguments{
\item{nsim}{number of realizations to simulate}

\item{paths}{Should population-level paths be returned?}

\item{observations}{Should simulated observations be returned? Requires that
a measurement process be defined in the stem object.}

\item{subject_paths}{Should population-level paths be mapped to subject-level
paths and returned (e.g. for the purpose of initializing a subject-level
collection of disease histories)? Only available for exact simulation via
the Gillespie direct method.}

\item{method}{either "gillespie" if simulating via Gillespie's direct method,
or "LNA" if simulating paths via the Linear Noise Approximation.}

\item{t0}{the time at which the system is initialized. If not supplied, it is
taken to be the first observation time if that is available, or otherwise
0. If supplied, it must be no later than the first observation time.}

\item{tmax}{the time at which simulation of the system is terminated. If not
supplied, defaults to the last observation time if not supplied.}

\item{timestep}{time discretization for smooth time-varying functions,
specifically seasonal terms, to be used in simulation when
method="gillespie". If not supplied, \code{timestep} will default to
\code{(tmax - t0)/50}.}

\item{census_times}{vector of times at which compartment counts should be
recorded. Required for \code{method = "LNA"}. If supplied and \code{method
= "gillespie"}, the compartment counts at census times are returned rather
than the full paths.}

\item{paths_as_array}{if TRUE, the simulated paths will be returned as an
array. Available only if \code{census_times} is specified.}

\item{datasets_as_array}{if TRUE, the simulated datasets will be returned as
an array.}

\item{messages}{should a message be printed when parsing the rates?}

\item{stem}{Stochastic epidemic model object. The dynamics slot must be
supplied at a minimum. If simulations from the measurement process are
desired, the measurement process slot must also be supplied.}
}
\value{
Returns a list with the simulated paths, subject-level paths, and/or
  datasets. If \code{paths = FALSE} and \code{observations = FALSE}, or if
  \code{paths = TRUE} and \code{observations = TRUE}, a list \code{nsim stem}
  paths and datasets, each returned either as a list or array, is returned.
  If \code{subject_paths = TRUE} and \code{method = "gillespie"}, a list of
  subject-level paths is also returned.

  If \code{paths = TRUE} and \code{observations = FALSE}, a list or array of
  simulated population-level paths is returned. If \code{method =
  "gillespie"} and \code{subject_paths = TRUE}, an additional list is
  returned, containing lists of subject-level mappings. Each sublist contains
  a vector of subject states at t0, along with a mapped path.

  If \code{paths = FALSE} and \code{observations = TRUE}, a list or array of
  simulated datasets is returned.
}
\description{
Simulations from a stochastic epidemic model.
}

