% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{stem}
\alias{stem}
\title{Construct a stem object.}
\usage{
stem(data = NULL, times = NULL, dynamics, parameters, constants = NULL,
  strata = NULL, timevar = "t", stem_settings = NULL, rmeasure = NULL,
  dmeasure = NULL, incidence_vars = NULL)
}
\arguments{
\item{data, times}{The observation times and data. The user may specify either
  a vector of observation times in the \code{times} argument, or a list of
  vectors of observation times, where the list element names identify the
  compartment_strata to which the observation times correspond (e.g.
  I_children, I_adults). \code{times} must be numeric and strictly
  increasing.

  \code{data} may be provided either as a data frame of observations with
  number of rows  equal to the number of observation times and with column
  names that identify the compartment_stratum (e.g. I_children, I_adults)
  being measured, or as a list of vectors, where the list element names
  identify the compartment_stratum and where each vector has a corresponding
  vector of observation times.}

\item{dynamics}{A list of objects describing the model dynamics, most
straighforwardly generated using the \code{stem_dynamics} function.}

\item{parameters}{Named vector of parameter values.}

\item{constants}{Names vector of constants. If constants are stratum
specific, they must be named appropriately - e.g. N_children, N_adults.}

\item{strata}{Names of strata. Must be specified if there are multiple
strata.}

\item{timevar}{name of the time variable, defaults to "t".}

\item{stem_settings}{otional list of simulation settings, most
straightforwardly generated using the \code{stem_control} function.}

\item{rmeasure, dmeasure}{functions to simulate from or evaluate the
likelihood of the measurement process. At least one must be specified.}

\item{incidence_vars}{vector of compartments for which the data consists of
incidence data.}
}
\value{
returns a \code{stem} list.
}
\description{
Construct a stem object.
}

