% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_dynamics.R
\name{stem_dynamics}
\alias{stem_dynamics}
\title{Generate the objects governing the dynamics of a stochastic epidemic model.}
\usage{
stem_dynamics(rates, parameters, state_initializer, compartments,
  tcovar = NULL, t0 = NULL, tmax = NULL, timestep = NULL,
  strata = NULL, constants = NULL, adjacency = NULL, messages = TRUE)
}
\arguments{
\item{rates}{list of rate lists, each generated by a call to the
\code{\link{rate}} function.}

\item{parameters}{numeric vector with named elements, the model parameters.}

\item{state_initializer}{list of initializer lists, each generated by a call
to \code{\link{stem_initializer}}. Used to initialize the system at the
first observation time.}

\item{compartments}{character vector of compartment names if there is a
single stratum, or if there are multiple strata a list of character vectors
where the name of each character vector is the compartment name and the
character vector lists the strata to in which the compartment exists. The
reserved word "ALL" can be used instead of listing all strata.}

\item{tcovar}{Matrix or data frame of time varying covariates, the first
column of which contains the times at which covariates change.}

\item{t0}{the time at which the system is initialized. If not supplied, it is
taken to be the first observation time if that is available, or otherwise
0. If supplied, it must be no later than the first observation time.}

\item{tmax}{the time at which simulation of the system is terminated. If not
supplied, defaults to the last observation time if not supplied.}

\item{timestep}{time discretization for smooth time-varying functions,
specifically seasonal terms, to be used in simulation when
method="gillespie". If not supplied, \code{timestep} will default to
\code{(tmax - t0)/50}.}

\item{strata}{vector of stratum names, required if the "ALL" reserved word is
  used

  (e.g. compartments = list(S = "ALL", I = "ALL", R = "ALL", D = "old");
  strata = c("infants", "young", "old");}

\item{constants}{optional. numeric vector with named elements that are
constants referenced in the rate functions.}

\item{adjacency}{optional matrix specifying the adjacency structure of
strata, with 0 entries indicating non-adjacency and 1 for adjacency. Rows
and columns must be labeled.}

\item{messages}{should a message be printed when parsing the rates?}
}
\value{
list with evaluated rate functions and objects for managing the
  bookkeeping for epidemic paths.
}
\description{
Generate the objects governing the dynamics of a stochastic epidemic model.
}

