% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference.R
\name{stem_inference}
\alias{stem_inference}
\title{Perform inference for a stochastic epidemic model using either the linear
noise approximation or agent-based Bayesian data augmentation.}
\usage{
stem_inference(stem_object, method, iterations, priors, kernel,
  t0_kernel = NULL, thin_params = 1,
  thin_latent_proc = ceiling(iterations/100), initialization_attempts = 500,
  messages = FALSE, monitor_MCMC = FALSE)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "lna" or "bda"}

\item{iterations}{number of iterations}

\item{priors}{Either a list of priors for model parameters, each generated by
a call to the \link{\code{prior}} function, or a list of three functions
supplied by the user with names "prior_density", "to_estimation_scale", and
"from_estimation_scale" (N.B. All three must be supplied). The first of
these functions should have two arguments that are numeric vectors of
parameters on their natural and estimation scales, specified in that order.
The functions for converting parameters to and from their estimation
scales. IMPORTANT NOTES: 1) The scale conversion functions must include the
dots argument, "...", in the function arguments. 2) The priors should not
include priors for the initial compartment counts or t0, which are
specified in the \code{\link{stem_initializer}} function and in the
\code{t0_kernel} argument.}

\item{kernel}{MCMC transition kernel generated by a call to the
\link{\code{kernel}} function. RWMH and adaptive RWMH are implemented for
jointly updating the model parameters (not including the initial
compartment counts which are handled separately).}

\item{t0_kernel}{output of \link{\code{t0_kernel}}, specifying the RWMH
transition kernel for t0 and the truncated normal distribution prior.}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100) so that every 100th path will be saved}

\item{messages}{should status messages be printed? defaults to TRUE.}

\item{monitor_MCMC}{should MCMC output be generated for the purposes of
monitoring convergence (defaults to FALSE)? If so, the acceptance rate and
the proposal covariance matrix (if adaptive RWMH is used) are printed
whenever messages are printed.}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Perform inference for a stochastic epidemic model using either the linear
noise approximation or agent-based Bayesian data augmentation.
}
