% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference.R
\name{stem_inference}
\alias{stem_inference}
\title{Perform inference for a stochastic epidemic model using either ordinary
differential equations, the linear noise approximation, or agent-based
Bayesian data augmentation.}
\usage{
stem_inference(stem_object, method, iterations, priors, mcmc_kernel,
  t0_kernel = NULL, thin_params = ceiling(iterations/1000),
  thin_latent_proc = ceiling(iterations/1000),
  initialization_attempts = 500, ess_args = NULL, print_progress = 0,
  messages = FALSE)
}
\arguments{
\item{stem_object}{a stochastic epidemic model object containing the dataset,
model dynamics, and measurement process.}

\item{method}{either "ode", "lna", or "bda".}

\item{iterations}{number of iterations}

\item{priors}{A list of three functions supplied by the user with names
"prior_density", "to_estimation_scale", and "from_estimation_scale" (N.B.
All three must be supplied). The first of these functions should have two
arguments that are numeric vectors of parameters on their natural and
estimation scales, specified in that order. The functions for converting
parameters to and from their estimation scales. IMPORTANT NOTES: 1) The
scale conversion functions must include the dots argument, "...", in the
function arguments. 2) The priors should not include priors for the initial
compartment counts or t0, which are specified in the
\code{stem_initializer} function and in the \code{t0_kernel} argument.}

\item{mcmc_kernel}{MCMC transition kernel generated by a call to the
\code{kernel} function.}

\item{t0_kernel}{output of \code{t0_kernel}, specifying the RWMH transition
kernel for t0 and the truncated normal distribution prior.}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100) so that every 100th path will be saved}

\item{initialization_attempts}{number of initialization attempts}

\item{ess_args}{list of elliptical slice sampling arguments}

\item{print_progress}{interval at which to print progress to a text file. If
0 (default) progress is not printed.}

\item{messages}{should status messages be printed? defaults to FALSE.}
}
\value{
list with posterior samples for the parameters and the latent
  process, along with MCMC diagnostics.
}
\description{
Perform inference for a stochastic epidemic model using either ordinary
differential equations, the linear noise approximation, or agent-based
Bayesian data augmentation.
}
