% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference_lna.R
\name{stem_inference_lna}
\alias{stem_inference_lna}
\title{Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.}
\usage{
stem_inference_lna(stem_object, iterations, prior_density, kernel,
  estimation_scales, thin_params, thin_latent_proc,
  initialization_attempts = 500, messages, monitor_MCMC)
}
\arguments{
\item{stem_object}{stochastic epidemic model object with model dynamics, the
measurement process, and a dataset.}

\item{iterations}{number of MCMC iterations}

\item{prior_density}{function to compute the prior density for the model
parameters}

\item{kernel}{list containing the MCMC kernel method, proposal covariance
matrix, and an external pointer for the compiled MCMC kernel function}

\item{estimation_scales}{character vector of estimation scales, each the
result of a}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100)}

\item{initialization_attempts}{number of attempts to initialize the latent
path before breaking.}

\item{messages}{should status messages be generated in an external text file?
If so, the iteration number is printed every 1000th iteration. defaults to
TRUE.}
}
\value{
list with parameter posterior samples and MCMC diagnostics
}
\description{
Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.
}

