% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_inference_lna.R
\name{stem_inference_lna}
\alias{stem_inference_lna}
\title{Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.}
\usage{
stem_inference_lna(stem_object, iterations, priors, mcmc_kernel, t0_kernel,
  thin_params, thin_latent_proc, initialization_attempts = 500,
  ess_args = NULL, messages)
}
\arguments{
\item{stem_object}{stochastic epidemic model object with model dynamics, the
measurement process, and a dataset.}

\item{iterations}{number of MCMC iterations}

\item{priors}{a list of named functions for computing the prior density as
well as transforming parameters to and from their estimation scales. The
functions should have the following names: "prior_density",
"to_estimation_scale", "from_estimation_scale". The prior_density function
must take two vectors as arguments, the model parameters (excluding initial
compartment volumes and t0) on their natural scales, and the model
parameters on their estimation scales. The functions for converting between
parameter scales should take vector of parameters as an argument, returning
a transformed vector (the function call has the form:
\code{transformed_vector <- conversion_function(original_vector)}).}

\item{mcmc_kernel}{list containing the mcmc_kernel method, proposal
covariance matrix, and an external pointer for the compiled mcmc_kernel
function}

\item{t0_kernel}{output of \code{t0_kernel}, specifying the RWMH transition
mcmc_kernel for t0 and the truncated normal distribution prior.}

\item{thin_params}{thinning interval for posterior parameter samples,
defaults to 1}

\item{thin_latent_proc}{thinning interval for latent paths, defaults to
ceiling(iterations/100)}

\item{initialization_attempts}{number of attempts to initialize the latent
path before breaking.}

\item{ess_args}{list of elliptical slice sampling settings, generated by a
call to \code{ess_settings}.}

\item{messages}{should status messages be generated in an external text file?
If so, the iteration number is printed whenever the latent process is
saved.}
}
\value{
list with parameter posterior samples and MCMC diagnostics
}
\description{
Approximate Bayesian inference for a stochastic epidemic model via the linear
noise approximation.
}
