% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_ode.R
\name{stem_ode}
\alias{stem_ode}
\title{Return the ODE path for a stochastic epidemic model.}
\usage{
stem_ode(stem_object, init_states, census_times)
}
\arguments{
\item{stem_object}{stochastic epidemic model object list}

\item{init_states}{vector of initial compartment counts}

\item{census_times}{times at which to evaluate the ode path}
}
\value{
matrix with the deterministic path evaluated at census times
}
\description{
Return the ODE path for a stochastic epidemic model.
}

