% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stem_ode_path}
\alias{stem_ode_path}
\title{Simulate a data matrix from the measurement process of a stochastic epidemic
model.}
\usage{
stem_ode_path(ode_times, census_times, census_inds, ode_pars, init_states,
  incidence_codes, ode_pointer, set_pars_pointer)
}
\arguments{
\item{ode_times}{vector of times at which the ODE must be evaluated}

\item{census_times}{vector of times at which the ODE should be censused}

\item{census_inds}{logical vector the same length as ode_times indicating
whether the ODE path is censused at each of the ODE times}

\item{ode_pars}{numeric matrix of parameters, constants, and time-varying
covariates at each of the ode_times}

\item{init_states}{matrix of initial state vectors}

\item{incidence_codes}{vector of incidence codes, or 0 if no incidence
compartments}

\item{ode_pointer}{external pointer to ODE integration fcn}

\item{set_pars_pointer}{external pointer to the function for setting the ODE
parameters.}
}
\value{
matrix containing the deterministic ODE path
}
\description{
Simulate a data matrix from the measurement process of a stochastic epidemic
model.
}

