% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_powers.R
\name{sub_powers}
\alias{sub_powers}
\title{Parse a string and substitute powers of the form a^b with pow(a,b).}
\usage{
sub_powers(string)
}
\arguments{
\item{string}{}
}
\value{
string suitable for C++
}
\description{
Parse a string and substitute powers of the form a^b with pow(a,b).
}
