% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpar.R
\name{tpar}
\alias{tpar}
\title{Generate a list to be used in specifying a time-varying parameter that has a
latent Gaussian distribution and is updated via elliptical slice sampling.}
\usage{
tpar(tparam_name, draws2par, times, n_draws, values = NULL)
}
\arguments{
\item{tparam_name}{name of the time--varying parameter}

\item{draws2par}{function for mapping a vector of N(0,1) draws of length
equal to the length of the \code{times} argument. The function should have
the following signature: \code{draws2par(draws, pars, path, tcovar, tparam, ...)}.
Here, draws is a vector of N(0,1) draws, pars is the vector of non-time
varying model hyperparameters (i.e. the parameters argument in
stem_dynamics), path is a matrix of latent process states, tcovar is a
matrix of time-varying covariates, and tparam is a matrix of time-varying
paramters. The function should return a vector of time-varying parameter
values.}

\item{times}{vector of times when the time-varying parameter changes.}

\item{n_draws}{number of N(0,1) random variates}

\item{values}{vector of values of N(0,1) draws for the time-varying
parameter, defaults to a vector of zeros. The values are computed by
applying the \code{draws2par} function to the supplied vector.}
}
\value{
list to be used in specifying a time-varying parameter. \describe{A
  time-varying parameter is defined as a (possibly non-linear) function of a
  set of N(0,1) draws that are updated via elliptical slice sampling.}
}
\description{
Generate a list to be used in specifying a time-varying parameter that has a
latent Gaussian distribution and is updated via elliptical slice sampling.
}
