% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpar.R
\name{tpar}
\alias{tpar}
\title{Generate a list to be used in specifying a time-varying parameter that has a 
latent Gaussian distribution and is updated via elliptical slice sampling.}
\usage{
tpar(tparam_name, times, draws2par)
}
\arguments{
\item{tparam_name}{name of the time--varying parameter}

\item{times}{vector of times when the time-varying parameter changes.}

\item{draws2par}{function for mapping a vector of N(0,1) draws of length 
equal to the length of the \code{times} argument. The function should take 
two arguments, the vector of model hyperparameters (i.e. the parameters 
argument) and a vector of N(0,1) draws, and return a vector of time-varying
parameter values. Note that the function should define a deterministic
transformation.}
}
\value{
list to be used in specifying a time-varying parameter. \describe{A 
  time-varying parameter is defined as a (possibly non-linear) function of a 
  set of N(0,1) draws that are updated via ellipeical slice sampling.}
}
\description{
Generate a list to be used in specifying a time-varying parameter that has a 
latent Gaussian distribution and is updated via elliptical slice sampling.
}
