% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_interval_widths}
\alias{update_interval_widths}
\title{Update factors and interval widths for automated factor slice sampling}
\usage{
update_interval_widths(interval_widths, n_expansions, n_contractions,
  n_expansions_c, n_contractions_c, slice_ratios, adaptation_factor)
}
\arguments{
\item{interval_widths}{vector of interval widths}

\item{n_expansions}{vector with number of expansion}

\item{n_contractions}{vector with number of contractions}

\item{n_expansions_c}{cumulative numbers of expansions}

\item{n_contractions_c}{cumulative numbers of contractions}

\item{slice_ratios}{vector for storing ratio of cumulative number of 
expansions over number of interval width changes}

\item{adaptation_factor}{}
}
\value{
adapt interval widths in place
}
\description{
Update factors and interval widths for automated factor slice sampling
}
