% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_interval_widths}
\alias{update_interval_widths}
\title{Update factors and interval widths for automated factor slice sampling}
\usage{
update_interval_widths(interval_widths, contraction_rates, adaptation_factor,
  target_contraction_rate)
}
\arguments{
\item{interval_widths}{vector of interval widths}

\item{contraction_rates}{rates of contractions along each slice direction}

\item{adaptation_factor}{weight for adaptation}

\item{target_contraction_rate}{expected number of contractions per iteration}
}
\value{
adapt interval widths in place
}
\description{
Update factors and interval widths for automated factor slice sampling
}
