% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lna_path.R
\name{update_lna_path}
\alias{update_lna_path}
\title{Sample a new LNA path via elliptical slice sampling.}
\usage{
update_lna_path(path_cur, lna_parameters, stoich_matrix, lna_ess_pointer,
  lna_times, lna_initdist_inds, param_update_inds, incidence_codes, data,
  measproc_indmat, obstime_inds, d_meas_pointer, parameters, constants,
  tcovar_censmat, do_census, do_incidence)
}
\arguments{
\item{path_cur}{list with the current LNA path along with its moments}

\item{lna_times}{vector of times at which the LNA must be evaluated}

\item{param_update_inds}{logical vector indicating at which of the times the
LNA parameters need to be updated.}

\item{data}{matrix containing the data}

\item{measproc_indmat}{matrix indicating which compartments are measured}

\item{obstime_inds}{list of indices indicating at which rows each measured
variable is recorded}

\item{d_meas_pointer}{external pointer to density function}

\item{parameters}{vector of model parameters}

\item{constants}{vector of constants}

\item{tcovar_censmat}{time-varying covariates at observation times}

\item{do_census}{Does the path need to be censused?}

\item{do_incidence}{Does incidence need to be computed?}
}
\value{
list with an updated LNA path along with its ODEs, the observed data
  log-likelihood, and the lna log-likelihood.
}
\description{
Sample a new LNA path via elliptical slice sampling.
}

