% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lna_path.R
\name{update_lna_path}
\alias{update_lna_path}
\title{Sample a new LNA path via elliptical slice sampling.}
\usage{
update_lna_path(path_cur, data, lna_parameters, propmat, pathmat, censusmat,
  emitmat, flow_matrix, lna_pointer, lna_set_pars_ptr, lna_times,
  lna_initdist_inds, param_update_inds, incidence_codes, census_incidence_codes,
  census_indices, measproc_indmat, obstime_inds, d_meas_pointer, parameters,
  constants, tcovar_censmat, do_prevalence, do_incidence, n_ess_updates)
}
\arguments{
\item{path_cur}{list with the current LNA path along with its ODE paths}

\item{data}{matrix containing the dataset}

\item{lna_parameters}{parameters, contants, time-varying covariates at LNA
times}

\item{propmat}{matrix in which to store the proposed N(0,1) draws}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{lna_pointer}{external LNA pointer}

\item{lna_times}{times at whicht eh LNA should be evaluated}

\item{lna_initdist_inds}{C++ column indices in the LNA parameter matrix for
the initial state}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{incidence_codes}{C++ indices for LNA count process compartments for
which incidence is to be computed}

\item{census_incidence_codes}{C++ indices for columns in the census matrix
containing incidence data}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{obstime_inds}{list of C++ indices indicating when each measurement
process is evaluated}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{parameters}{vector of model parameters}

\item{constants}{vector of model constants}

\item{tcovar_censmat}{time-varying covariates at LNA times}

\item{do_prevalence}{should prevalence be computed?}

\item{do_incidence}{should incidence be computed?}
}
\value{
list with an updated LNA path along with its stochastic
  perturbations, the observed data log-likelihood, and the lna
  log-likelihood, and a record of the number of elliptical slice sampling
  proposals
}
\description{
Sample a new LNA path via elliptical slice sampling.
}
