% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lna_path.R
\name{update_lna_path}
\alias{update_lna_path}
\title{Sample a new LNA path via elliptical slice sampling.}
\usage{
update_lna_path(
  path_cur,
  data,
  lna_parameters,
  lna_param_vec,
  init_volumes_cur,
  init_volumes_prop,
  initdist_objects,
  tparam,
  pathmat_prop,
  censusmat,
  draws_prop,
  ess_draws_prop,
  emitmat,
  flow_matrix,
  stoich_matrix,
  lna_times,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  lna_param_inds,
  lna_const_inds,
  lna_tcovar_inds,
  lna_initdist_inds,
  param_update_inds,
  census_indices,
  lna_event_inds,
  measproc_indmat,
  svd_d,
  svd_U,
  svd_V,
  lna_pointer,
  lna_set_pars_pointer,
  d_meas_pointer,
  do_prevalence,
  n_ess_updates,
  ess_schedule,
  lna_bracket_width,
  joint_tparam_update,
  joint_initdist_update,
  step_size
)
}
\arguments{
\item{path_cur}{list with the current LNA path along with its ODE paths}

\item{data}{matrix containing the dataset}

\item{tparam}{list of time-varying parameters}

\item{censusmat}{template matrix for the LNA path and incidence at the
observation times}

\item{emitmat}{matrix in which to store the log-emission probabilities}

\item{stoich_matrix}{LNA stoichiometry matrix}

\item{forcing_inds}{logical vector of indicating at which times in the
time-varying covariance matrix a forcing is applied.}

\item{param_update_inds}{logical vector indicating when to update the
parameters}

\item{census_indices}{C++ row indices of LNA times when the path is to be
censused}

\item{measproc_indmat}{logical matrix for evaluating the measuement process}

\item{svd_d, svd_U, svd_V}{objects for computing the SVD of LNA
diffusion matrics}

\item{d_meas_pointer}{external pointer for the measurement process function}

\item{do_prevalence}{should prevalence be computed?}

\item{n_ess_updates}{number of elliptical slice sampling updates}

\item{step_size}{initial step size for the ODE solver (adapted internally,
but too large of an initial step can lead to failure in stiff systems).}

\item{tparam_update}{if TRUE then time-varying parameters are updated jointly
along with the LNA path}
}
\value{
list with an updated LNA path along with its stochastic
  perturbations, the observed data log-likelihood, and the lna
  log-likelihood, and a record of the number of elliptical slice sampling
  proposals
}
\description{
Sample a new LNA path via elliptical slice sampling.
}
