% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_princomps}
\alias{update_princomps}
\title{Update principal component Metropolis eigenvectors and eigenvalues}
\usage{
update_princomps(eigenvalues, eigenvectors, comp_weights, kernel_cov,
  update_weights)
}
\arguments{
\item{eigenvalues}{vector of eigenvalues of the proposal covariance matrix}

\item{eigenvectors}{matrix of eigenvectors for the proposal covariance}

\item{comp_weights}{vector of component weights}

\item{kernel_cov}{proposal covariance matrix}

\item{update_weights}{should the weights for the direction sampling 
distribution be recomputed.}
}
\value{
update eigenvalues and eigenvectors in place
}
\description{
Update principal component Metropolis eigenvectors and eigenvalues
}
